### 2021-09-10 #132 @whedon
[PRE REVIEW]: A virtual training module for introducing the use of Amazon Web Services
**Submitting author:** @abhijna (<a href="http://orcid.org/0000-0002-5861-3737">Abhijna Parigi</a>)
**Repository:** <a href="https://github.com/nih-cfde/training-and-engagement" target ="_blank">https://github.com/nih-cfde/training-and-engagement</a>
**Version:** v1.0.0
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
