### 2021-09-11 #132 @erahulkulkarni
Good Day,

Sorry, but I have very little experience in the use of Amazon Web Services.

So in one way, the readability of the documentation should help me get
started with Amazon Web Services.

But since I have little experience I would not be able to suggest other /
better ways of usage.

Please let me know if it would still be okay if I agree to review?

Thanks and Regards,
*Rahul Kulkarni*

On Fri, Sep 10, 2021 at 11:02 PM Abhijna Parigi ***@***.***>
wrote:

> Suggested reviewers: Rahul Kulkarni, Jason Williams, Jens Lechtenbörger,
> Luke Johnston
>
