### 2021-09-17 #132 @jgorzalski
Hello @abhijna! Sorry for the slow reply, we've been a little backlogged at JOSE due to the pandemic and are trying to keep up with everything. I've read through your paper and I need a little clarification before we move this forward. At JOSE we are looking to publish modules and course lessons with their own objectives and activities that could be taught over three to five weeks of a semester. I'm getting the sense from this paper that the information and tools presented here amount to more of a two hour training session or a one time in class activity. Am I understanding that correctly? Thank you for the clarification.
