### 2021-09-20 #132 @abhijna
Hello @jgorzalski, that is correct. We offer a 2-hour, one-time virtual lesson that provides a hands-on introduction to configuring, launching, and connecting to an AWS instance. The lesson is intended for novice users who wish to learn about AWS (and cloud computing) without the commitment of a semester-long course. Our workshops are widely successful and attract diverse audiences (e.g. professors, students, and non-academic professionals). Please let me know if you need further clarification.
