### 2021-11-13 #132 @labarba
@giuliaguidi 👋  — We received this submission to The Journal of Open Source Software, JOSE: **A virtual training module for introducing the use of Amazon Web Services**. Could you help us out with a review?
JOSE is a fully volunteer-led, open journal for publishing open source software for educational uses and open learning modules or tutorials. Have a look at our [docs](https://openjournals.readthedocs.io/en/jose/).
