### 2021-11-15 #132 @giuliaguidi
Hi,

I would be happy to help, please let me know how to proceed.

Best,
Giulia Guidi

On Sat, Nov 13, 2021 at 10:52 AM Lorena A. Barba ***@***.***>
wrote:

> @giuliaguidi <https://github.com/giuliaguidi> 👋 — We received this
> submission to The Journal of Open Source Software, JOSE: *A virtual
> training module for introducing the use of Amazon Web Services*. Could
> you help us out with a review?
> JOSE is a fully volunteer-led, open journal for publishing open source
> software for educational uses and open learning modules or tutorials. Have
> a look at our docs <https://openjournals.readthedocs.io/en/jose/>.
>
