### 2021-11-15 #132 @willingc
Hi @giuliaguidi, The best place to begin learning about the review process is to start in our [reviewer guide](https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html). Please feel free to reach out with questions. Thanks. :sunny:
