### 2021-11-15 #132 @giuliaguidi
Great, I'll have a look later this week. Is there a deadline for the review?

Best,
Giulia

On Mon, Nov 15, 2021 at 3:40 PM Carol Willing ***@***.***>
wrote:

> Hi @giuliaguidi <https://github.com/giuliaguidi>, The best place to begin
> learning about the review process is to start in our reviewer guide
> <https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html>.
> Please feel free to reach out with questions. Thanks. ☀️
>
