### 2022-01-14 #132 @giuliaguidi
I'll complete it by the end of the next week.

Best,
Giulia

On Thu, Jan 13, 2022 at 12:50 PM Abhijna Parigi ***@***.***>
wrote:

> Hello @willingc <https://github.com/willingc> and @giuliaguidi
> <https://github.com/giuliaguidi>, are there any updates on the review
> status of this paper? Please let me know if you need anything.
>
