### 2022-01-14 #132 @labarba
Hi everybody — the handling editor, @willingc, will need to add one more reviewer here before running `@whedon start review`, which will open a new issue thread that will show a review checklist for each reviewer (and close this thread).

@giuliaguidi : You can start reviewing the materials, but the conversation about revisions, etc., will happen in the Review thread.
