+++ hackmd/UNIX.md
@@ -1,23 +1,31 @@
-# An Introduction to UNIX for Cloud Computing 
-
-**When:**  
-
-**Instructors:**  
-
-**Helpers:**   
-
-**Zoom:**  
-
-**Slides:**  
-
-**Computing environment:** https://mybinder.org/v2/gh/nih-cfde/training-rstudio-binder/data?urlpath=rstudio
-
-**Cheatsheet:** https://training.nih-cfde.org/en/latest/General-Tools/Cheat-Sheets/bash_cheatsheet/
+# An Introduction to UNIX for Remote Computing: January 19, 2022
 
 
-#### Workshop Overview  
+**When:** Wednesday, January 19, 2022 from 10:00 am-12:00 pm PDT
 
-This free 2-hour workshop introduces the UNIX command line. It was designed for scientists and clinicians who need to use cloud-based and remote computers for basic and biomedical research. 
+**Instructors:** Dr. Rayna Harris
+
+**Helpers:**  Jessica Lumian and Jeremy Walter
+
+**Connect via**  [![Zoom](https://img.shields.io/badge/Zoom-2D8CFF?style=for-the-badge&logo=zoom&logoColor=white)](https://zoom.us/j/7575820324?pwd=d2UyMEhYZGNiV3kyUFpUL1EwQmthQT09)
+
+
+**Computing environment:** [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/nih-cfde/training-rstudio-binder/data?urlpath=rstudio)
+
+:::warning
+#### Questions: 
+
+1. What city/state/country are you in now?
+1. What time is it?
+1. Why did you hope to learn from this workshop? 
+:::
+
+
+#### Lesson Overview  
+
+This lesson introduces the UNIX command line to scientists and clinicians who need to use cloud-based and remote computers for basic and biomedical research. 
+
+These commands used in today's workshop are commonly used by practitioners across disciplines. By practicing the commands, you will gain familiarity with the command line and confidence as a programmer. These commands can be used on a local or remote computer. We hope that you notice an improvement in the speed and reproducibility of your bioinformatics workflow. 
 
 :::success
 #### Today's Learning Goals
@@ -25,25 +33,12 @@ This free 2-hour workshop introduces the UNIX command line. It was designed for 
 * Understand basic UNIX command structure 
 * Navigate through hierarchical directory structures
 * Read, write, create, copy, move, and remove files
-* Understand wildcard expression 
-* Combine commands into workflows
+* Understand wildcards and regular expression 
+* Redirect outputs and write for loops to build reproducible workflows
 :::
 
-The lesson materials were adapted from the UC Davis Data Lab's [Intro to Cloud Computing](https://ngs-docs.github.io/2021-august-remote-computing/) workshop, Data Carpentry's [Introduction to the Command Line for Genomics](https://datacarpentry.org/shell-genomics/) lesson, and the Lab for Data Intensive Biology's [Advanced Beginner/Intermediate Shell](https://dib-training.readthedocs.io/en/pub/2016-01-13-adv-beg-shell.html) workshop. Today's workshop is designed to prepare you for taking subsequent cloud computing and bioinformatic workshops. 
-
-
-:::info
-While we wait to get started --
-
-1. :heavy_check_mark: Have you looked at the [pre-workshop resources page](https://github.com/nih-cfde/training-and-engagement/wiki/Resources-for-Workshop-Attendees)?
-
-2. :pencil: Please fill out our pre-workshop survey if you have not already done so:
-[Click here!](https://forms.gle/kf87XXWa7cZFDXUA9)
-
-
[…]
