### 2022-01-20 1c48e3a
final november version, dates removed
