### 2022-01-20 91f304a
jan workshop final, dates and whitespace removed
