+++ hackmd/UNIX.md
@@ -1,14 +1,12 @@
-# An Introduction to UNIX for Remote Computing: January 19, 2022
+# An Introduction to UNIX for Remote Computing 
 
+**When:**  
 
-**When:** Wednesday, January 19, 2022 from 10:00 am-12:00 pm PDT
+**Instructors:**  
 
-**Instructors:** Dr. Rayna Harris
+**Helpers:**   
 
-**Helpers:**  Jessica Lumian and Jeremy Walter
-
-**Connect via**  [![Zoom](https://img.shields.io/badge/Zoom-2D8CFF?style=for-the-badge&logo=zoom&logoColor=white)](https://zoom.us/j/7575820324?pwd=d2UyMEhYZGNiV3kyUFpUL1EwQmthQT09)
-
+**Connect via**  
 
 **Computing environment:** [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/nih-cfde/training-rstudio-binder/data?urlpath=rstudio)
 
@@ -19,7 +17,6 @@
 1. What time is it?
 1. Why did you hope to learn from this workshop? 
 :::
-
 
 #### Lesson Overview  
 
@@ -43,7 +40,6 @@ Please refer to our [UNIX Cheatsheet](https://training.nih-cfde.org/en/latest/Ge
 #### Table of Contents
 [TOC]
 
-
 ---
 
 ## 1. The shell and terminal
@@ -56,9 +52,7 @@ For this remote workshop, we will be using a custom-created computing environmen
 
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/nih-cfde/training-rstudio-binder/data?urlpath=rstudio)
 
-
 ![](https://hackmd.io/_uploads/HkZKMYfTY.png)
-
 
 To help with readability, type `PS1='$ '` in the terminal to remove the very long computer name from the path. Then type `clear` to clear the screen of irrelevant warning messages. 
 
@@ -149,7 +143,6 @@ drwxr-xr-x 2 jovyan jovyan   4096 Jan 18 21:13 seattle
 drwxr-xr-x 2 jovyan jovyan   4096 Jan 18 21:13 southpark
 ```
 
-
 Flags (sometimes called options) allow us to finely control the behavior of the command. But how did we know to add `-l` after ls? The [`ls` manual ](https://man7.org/linux/man-pages/man1/ls.1.html) describes the command and all its options in detail. Like most commands, you can type the command followed `--help` to view the manual in your terminal.
 
 ```
@@ -162,11 +155,9 @@ ls --help
 
 You can use multiple flags, wildcards, and specify directories to modify the behavior of a command. What does the command `ls` do when used with the following option:
 
-
 1. `ls -a`
 2. `ls -F`
 3. `ls -aF`
-
 
 :::spoiler
 
@@ -181,11 +172,8 @@ Now we have seen how to list the contents of folders on our computers and what i
 To move from one directory to the other, we use the `cd` command to **change directories**. We can use the `pwd` and/or `ls` commands to confirm that we did indeed change directories.  Because you can change directories using either the relative or absolute path, there are multiple ways to successfully move up or down in the directory hierarchy.
 
 Let's return to our home directory using the `cd` command and a relative path, then print the working directory to confirm.  
-
  
 Let's practice using the cd and ls commands to explore files in different directories.  
- 
-
 
 Because books/ is in our working directory, we can navigate there with a relative path. What files are in the `books` directory and how large are they?
 
@@ -207,7 +195,6 @@ We can see the following files:
[…]
