### 2022-01-20 51c3ce2
Dec 1 2020 workshop notes
