+++ hackmd/AWS.md
@@ -1,78 +1,301 @@
-# A Hands On Introduction to AWS 
+# A Hands On Introduction to AWS:
 
 **When:**  
 
 **Instructors:**  
 
+**Moderator:**  
+
 **Helpers:**  
 
+
+:speech_balloon: Please fill out our [pre-workshop survey](https://docs.google.com/forms/d/e/1FAIpQLSd7SuaESZ0g8MZswdgxlnjC6m8r5fQUcyH6QYMzN2RW4mjCew/viewform)!
+
+
 ## Description
-This 2 hour hands on tutorial will introduce you to creating a computer "in the cloud" and logging into it, via Amazon Web Services. We'll create a small general-purpose Linux computer connect to it, and run a small job, while discussing the concepts and technologies involved.
 
-## Agenda
+This 2 hour hands-on tutorial will introduce you to creating a computer "in the cloud" and logging into it, via Amazon Web Services. We'll create a small general-purpose Linux computer, connect to it, and run a small job while discussing the concepts and technologies involved.
 
-### Introduction and workshop goals
+:computer: Links for moderator to share with participants:
 
-Hello!
-I'm Titus Brown, a professor here at UC Davis. I'm joined today by Dr. Abhijna Parigi. We are both part of the training and engagement team for the [NIH Common Fund Data Ecosystem](https://nih-cfde.org/), a project supported by the NIH to increase data reuse and cloud computing for biomedical research.
-You can contact us at ctbrown@ucdavis.edu, and abhijna.parigi@gmail.com.
-This is our first workshop for our Amazon Web services lesson, and we have the following goals:
+- [x] [Pre-workshop survey](https://docs.google.com/forms/d/e/1FAIpQLSd7SuaESZ0g8MZswdgxlnjC6m8r5fQUcyH6QYMzN2RW4mjCew/viewform)
+- [x] [Workshop notes link](https://hackmd.io/4zmi-AQBQIelVp3_WFaxyQ?view)
+- [x] [Post-workshop survey](https://docs.google.com/forms/d/e/1FAIpQLSe2qR2t3PIJlRWoOERPK-lRz3JvQv6z_nHGCCQ1771-EeZ0EQ/viewform)
+
+## Hello!
+
+Your instructors are part of the training and engagement team for the [NIH Common Fund Data Ecosystem](https://nih-cfde.org/), a project supported by the NIH to increase data reuse and cloud computing for biomedical research.
+
+
+This is only our second workshop for our Amazon Web services lesson :slightly_smiling_face:	, and we have the following goals:
 * run the material by you all!
 * gather questions and refine the tutorial materials!
 * help you think about if and how to use cloud computers for your work!
+
 So, please ask lots of questions, and even the ones we can't answer yet we'll figure out for you!
+
+### Costs and payment
+
 Today, everything you do will be paid for by us (well, the NIH). In the future, if you create your own AWS account, you'll have to put your own credit card on it. We'd be happy to answer questions about how to pay for AWS - we've used invoicing as well as credit cards.
-How we will do this!
-* first, an instructor will walk through process, taking questions
-* then everyone will do it themselves!
-* then an instructor will walk through things again, taking more questions.
-* if you have questions, drop them in chat!
 
-### Some background
-Amazon and Google, among others, rent compute resources for money.
+:smiley_cat: **Your free login credentials will work for the next 24 hours**
+
+### Workshop structure and plan
+
+* Brief introduction to AWS and the cloud
+* An instructor will do a demo, taking questions
+* Everyone will do it themselves!
+* Install and run things in the cloud computer
+* Take your questions
+
+### How to ask questions
+If you have questions at any point, 
+- Drop them in the chat, or
+- Unmute yourself and ask during the workshop, or
[…]
