+++ hackmd/AWS.md
@@ -1,4 +1,4 @@
-# A Hands On Introduction to AWS:
+# A Hands On Introduction to AWS
 
 **When:**  
 
@@ -8,26 +8,27 @@
 
 **Helpers:**  
 
-
-:speech_balloon: Please fill out our [pre-workshop survey](https://docs.google.com/forms/d/e/1FAIpQLSd7SuaESZ0g8MZswdgxlnjC6m8r5fQUcyH6QYMzN2RW4mjCew/viewform)!
+ - [ ] Please fill out our [pre-workshop survey](https://forms.gle/TNhgV1KYCHgRKr1v6)!
 
 
 ## Description
 
 This 2 hour hands-on tutorial will introduce you to creating a computer "in the cloud" and logging into it, via Amazon Web Services. We'll create a small general-purpose Linux computer, connect to it, and run a small job while discussing the concepts and technologies involved.
 
-:computer: Links for moderator to share with participants:
+Links for moderator to share with participants:
 
-- [x] [Pre-workshop survey](https://docs.google.com/forms/d/e/1FAIpQLSd7SuaESZ0g8MZswdgxlnjC6m8r5fQUcyH6QYMzN2RW4mjCew/viewform)
-- [x] [Workshop notes link](https://hackmd.io/4zmi-AQBQIelVp3_WFaxyQ?view)
-- [x] [Post-workshop survey](https://docs.google.com/forms/d/e/1FAIpQLSe2qR2t3PIJlRWoOERPK-lRz3JvQv6z_nHGCCQ1771-EeZ0EQ/viewform)
+- [ ] [Pre-workshop survey](https://forms.gle/TNhgV1KYCHgRKr1v6)
+- [ ] [Workshop notes link](https://hackmd.io/jTqsKMSZRU-L6NkJM4fZew?view)
+- [ ] [mobaXterm install link](https://mobaxterm.mobatek.net/)
+- [ ] [AWS lessons](https://training.nih-cfde.org/en/latest/Bioinformatics-Skills/Introduction_to_Amazon_Web_Services/introtoaws1/) 
+- [ ] [Post-workshop survey](https://forms.gle/2tGthbddRCQ72vQG9)
+
 
 ## Hello!
 
 Your instructors are part of the training and engagement team for the [NIH Common Fund Data Ecosystem](https://nih-cfde.org/), a project supported by the NIH to increase data reuse and cloud computing for biomedical research.
 
-
-This is only our second workshop for our Amazon Web services lesson :slightly_smiling_face:	, and we have the following goals:
+This is our third workshop for our Amazon Web services lesson :slightly_smiling_face:, and we have the following goals:
 * run the material by you all!
 * gather questions and refine the tutorial materials!
 * help you think about if and how to use cloud computers for your work!
@@ -38,23 +39,26 @@ So, please ask lots of questions, and even the ones we can't answer yet we'll fi
 
 Today, everything you do will be paid for by us (well, the NIH). In the future, if you create your own AWS account, you'll have to put your own credit card on it. We'd be happy to answer questions about how to pay for AWS - we've used invoicing as well as credit cards.
 
-:smiley_cat: **Your free login credentials will work for the next 24 hours**
+
+ _**Your free login credentials will work for the next 24 hours !**_
+
 
 ### Workshop structure and plan
 
 * Brief introduction to AWS and the cloud
-* An instructor will do a demo, taking questions
-* Everyone will do it themselves!
+* Set up an instance and connect to it
 * Install and run things in the cloud computer
+* Learn how to download output files to local machine
 * Take your questions
 
 ### How to ask questions
 If you have questions at any point, 
 - Drop them in the chat, or
[…]
