+++ hackmd/AWS.md
@@ -15,32 +15,39 @@
 
 This 2 hour hands-on tutorial will introduce you to creating a computer "in the cloud" and logging into it, via Amazon Web Services. We'll create a small general-purpose Linux computer, connect to it, and run a small job while discussing the concepts and technologies involved.
 
-Links for moderator to share with participants:
 
-- [ ] [Pre-workshop survey](https://forms.gle/TNhgV1KYCHgRKr1v6)
-- [ ] [Workshop notes link](https://hackmd.io/jTqsKMSZRU-L6NkJM4fZew?view)
-- [ ] [mobaXterm install link](https://mobaxterm.mobatek.net/)
-- [ ] [AWS lessons](https://training.nih-cfde.org/en/latest/Bioinformatics-Skills/Introduction_to_Amazon_Web_Services/introtoaws1/) 
-- [ ] [Post-workshop survey](https://forms.gle/2tGthbddRCQ72vQG9)
+While we wait to get started --
+
+1. :heavy_check_mark: Have you checked out the [pre-workshop resources page](https://github.com/nih-cfde/training-and-engagement/wiki/Resources-for-Workshop-Attendees)?
+
+2. If you are on a windows computer, make sure you have Mobax term: https://mobaxterm.mobatek.net/
+
+
 
 
 ## Hello!
 
-Your instructors are part of the training and engagement team for the [NIH Common Fund Data Ecosystem](https://nih-cfde.org/), a project supported by the NIH to increase data reuse and cloud computing for biomedical research.
+Your instructors are both part of the training and engagement team for the [NIH Common Fund Data Ecosystem](https://nih-cfde.org/), a project supported by the NIH to increase data reuse and cloud computing for biomedical research.
 
-This is our third workshop for our Amazon Web services lesson :slightly_smiling_face:, and we have the following goals:
-* run the material by you all!
-* gather questions and refine the tutorial materials!
-* help you think about if and how to use cloud computers for your work!
+
+**Have you heard of the NIH Common Fund Data Ecosystem?**
+
+Put up a :heavy_check_mark: for yes and a :negative_squared_cross_mark: for no!
+
+We have the following goals for this workshop:
+
+* Help you think about if and how to use cloud computers for your work!
+* Gather questions, feedback and refine the tutorial materials!
 
 So, please ask lots of questions, and even the ones we can't answer yet we'll figure out for you!
 
 ### Costs and payment
 
-Today, everything you do will be paid for by us (well, the NIH). In the future, if you create your own AWS account, you'll have to put your own credit card on it. We'd be happy to answer questions about how to pay for AWS - we've used invoicing as well as credit cards.
+Today, everything you do will be paid for by us. In the future, if you create your own AWS account, you'll have to put your own credit card on it. We'd be happy to answer questions about how to pay for AWS.
 
 
- _**Your free login credentials will work for the next 24 hours !**_
+
+:smiley_cat: Your free login credentials will work for the next 8 hours
 
 
 ### Workshop structure and plan
@@ -54,16 +61,17 @@ Today, everything you do will be paid for by us (well, the NIH). In the future, 
 ### How to ask questions
 If you have questions at any point, 
 - Drop them in the chat, or
-- Direct messages to the moderator (Marisa Lim) are welcome, or
-- Unmute yourself and ask during the workshop, or
-- Use the raise hand reaction in zoom
+- Direct messages to the moderator (Saranya Canchi) are welcome, or
+- Unmute yourself and ask during the workshop
 
[…]
