+++ hackmd/AWS.md
@@ -1,4 +1,4 @@
-# A Hands On Introduction to AWS
+# A Hands-on Introduction to AWS
 
 **When:**  
 
@@ -8,26 +8,25 @@
 
 **Helpers:**  
 
+
 ## Description
 
 This 2 hour hands-on tutorial will introduce you to creating a computer "in the cloud" and logging into it, via Amazon Web Services. We'll create a small general-purpose Linux computer, connect to it, and run a small job while discussing the concepts and technologies involved.
 
 
-While we wait to get started --
-
-1. :heavy_check_mark: Have you checked out the [pre-workshop resources page](https://github.com/nih-cfde/training-and-engagement/wiki/Resources-for-Workshop-Attendees)?
-
-2. If you are on a windows computer, make sure you have installed MobaXterm: https://mobaxterm.mobatek.net/
+> :smiley_cat: While we wait to get started
+> 1. :heavy_check_mark: Have you checked out the [pre-workshop resources page](https://github.com/nih-cfde/training-and-engagement/wiki/Resources-for-Workshop-Attendees)?
+> 2. :pencil: Please fill out our pre-workshop survey if you have not already done so! [Please click this link](https://forms.gle/hCtbXrDyH9omAYyB6) 
+> 3. If you are on a windows computer, make sure you have Mobaxterm. [Check out the install guide!](https://hackmd.io/YivVGhznRwqVO3fxdGpkOQ?view)
 
 
 ## Hello!
 
-Your instructors are both part of the training and engagement team for the [NIH Common Fund Data Ecosystem](https://nih-cfde.org/), a project supported by the NIH to increase data reuse and cloud computing for biomedical research.
+Your instructors and helpers are part of the training and engagement team for the [NIH Common Fund Data Ecosystem](https://nih-cfde.org/), a project supported by the NIH to increase data reuse and cloud computing for biomedical research.
 
 
-**Have you heard of the NIH Common Fund Data Ecosystem?**
-
-Put up a :heavy_check_mark: for yes and a :negative_squared_cross_mark: for no!
+> 👍 **Have you heard of the NIH Common Fund Data Ecosystem?**
+> Put up a :heavy_check_mark: for yes and a :negative_squared_cross_mark: for no!
 
 We have the following goals for this workshop:
 
@@ -40,36 +39,33 @@ So, please ask lots of questions, and even the ones we can't answer yet we'll fi
 
 Today, everything you do will be paid for by us. In the future, if you create your own AWS account, you'll have to put your own credit card on it. We'd be happy to answer questions about how to pay for AWS.
 
-
-
-:smiley_cat: Your free login credentials will work for the next 8 hours
-
+> :smiley_cat: Your free login credentials will work for the next 24 hours
 
 ### Workshop structure and plan
 
-* Brief introduction to AWS and the cloud
-* Set up an instance and connect to it
-* Install and run things in the cloud computer
-* Learn how to download output files to local machine
-* Take your questions
+- [ ]  Brief introduction to AWS and the cloud
+- [ ]  Set up an instance and connect to it
+- [ ]  Install and run things in the cloud computer
+- [ ]  Learn how to download output files to local machine
+- [ ]  Take your questions
 
 ### How to ask questions
 If you have questions at any point, 
 - Drop them in the chat, or
-- Direct messages to the moderator (Saranya Canchi) are welcome, or
+- Direct messages to the moderator are welcome, or
 - Unmute yourself and ask during the workshop
 
-We're going to use the "raise hand" reaction in zoom to make sure people are on board during the hands-on activities.
[…]
