+++ docs/General-Tools/UNIX/unix_2.md
@@ -1,185 +1,190 @@
-# Navigating directories
+# Navigating files and directories
 
 UNIX commands are like sentences that can be very simple or complex. The simplest commands consist of only the command name. Many require the name of a file or directory and allow specially formatted arguments, known as flags or options, which modify the default behavior of the program. The grammar of a shell allows you to combine existing tools into powerful pipelines and handle large volumes of data automatically. Sequences of commands can be written into a script, improving the reproducibility of workflows. The ease of getting things done via the shell will increase with your exposure to the program.
 
-
 We should note that _folders_ are called **directories** at the command line. For all intents and purposes, they can be used interchangeably, but if you'd like more information please read about ["the folder metaphor"](https://en.wikipedia.org/wiki/Directory_%28computing%29#Folder_metaphor).
 
-![](https://i.imgur.com/tS4uw77.png)
-
-
-This Binder comes preloaded with data provided by your instructors.  _If you want to do these exercises locally, on your own computer, you can [download the data here](https://s3-us-west-1.amazonaws.com/dib-training.ucdavis.edu/shell-data.zip)._
-
-For today's lesson, we will focus on three different sets of data. `2cities` contains a compressed text file containing the book A Tale of Two Cities, `SouthParkData` contains a compressed csv file containing all the lines spoken by each character across 14 seasons. This dataset is useful for teaching UNIX commans on medium sized text data. The `data/MiSeq` directory contains FASTQ and FASTA files that are commonly used in next-generation sequencing experiments. These data are useful for teaching commonly used UNIX commands for exploring genome-scale data. 
-
-
+This Binder comes preloaded with data provided by your instructors.  _If you want to do these exercises locally, on your own computer, you can [download the data here](https://s3.us-west-1.amazonaws.com/dib-training.ucdavis.edu/shell-data2.zip)._
 
 The commands `pwd` and `ls` are two simple commands that can be used to answer the two commonly asked questions "where am I?" and "what files are here?". We will use these frequently through the next sections.
 
 To answer the question "where am I?", we can use the **print working directory** or `pwd` command to see what directory we are currently located in. 
 
+=== "Input"
+	
+	```
+	pwd
+	```
 
-```
-pwd
-```
-
-
-This will print **absolute path** to the directory where we are located. An absolute path shows the complete series of directories you need to locate either a directory or a file starting from the **root directory** of your computer. The absolute path to the root directory is `/`. A useful way to start thinking about directories and files is through levels. At the highest level of your computer, you have the root directory. Everything that is contained in your computer is located in directories below your root directory. 
-
-
-The **home directory** is typically two levels down. For many mac users, the home directory is `/Users/USERNAME`. If you are using the Binder provided for this workshop, the home directory is `/home/jovyan`. Because the absolute path to the home directory is different for every user, you can refer to the home directory with the tilde symbol `~`.
+=== "Expected Output"
 
 
-```
-/home/jovyan
-```
+	```
+	/home/jovyan
+	```
 
 
[…]
