+++ docs/General-Tools/UNIX/unix_4.md
@@ -1,55 +1,90 @@
-# Building pipelines
+# Creating and modifying files
 
-By default, the output of many UNIX commands is sent to the
-standard output, or "stdout". (Some commands also produce a standard error, or "stderr", which is where errors are printed; and standard input, or "stdin", which is where input comes from.)
+We are used to copying and moving files using a GUI. These functions can be also carried out at the command line.
 
-Much of the power of the UNIX command line comes from working with
-stdout output, and if you work with UNIX a lot, you'll see characters
-like the `|` (pipe), `>` (redirect), `>>` (append)  thrown around. These
-are redirection commands that say, respectively, "send stdout to a new
-file", "append stdout to an existing file", and "send stdout from one
-program to another program's stdin."
-
-If you know you want to save an output file, you can use the redirect symbol `>`. Note, if you want to save a file in a different directory, that directory must exist.
-
-Let's practice building pipelines using different datasets
-
-=== "books"
+The `cp` and `mv` commands can be used to copy and move (or rename) files and directories respectively. For both commands, you must specify the old and new names. Specifying the path is necessary if you want to move files out of the current working directory.
 
 ```
-cd ~/books
-wc -l *txt
-wc -l *txt | sort
-wc -l *txt | sort -nr
-wc -l *txt | sort -nr
-mkdir results
-wc -l *txt | sort -nr > results/booklengths.txt
-
+cp [original-filename] [copy-filename]
+mv [original-filename] [new-filename]
 ```
 
-
-
-=== "MiSeq"
-
+Let's make a copy of some raw data before we start modifying it. 
 
 ```
-cd ~/MiSesq
-for file in *fastq
-do
-echo $file
-grep CATTAG $file | wc -l
-done
+cp book.txt book-copy.txt
+ls
 ```
 
+The `mv` command can be used to either move files to a new location or to rename them (which is essentially moving the contents from the old filename to the new file name. Let's use the `mv` command to rename one of the books.
 
-=== "southpark"
+```
+mv A-tale-of-two-cities.txt Two-Cities.txt
+```
+
+If you want to move your book-copy.txt to a new folder, you first have to create that folder using the command `mkdir`. Then you move the copy. We can use `ls *` to list files in this directory and subdirectories.
+
+```
+mkdir book-copies
+mv book-copy.txt book-copies
+ls *
+```
+
+Take care when naming and renaming files. File names should not contain spaces or slashes. The use of capital letters (e.g. CamelCase) and underscores (e.g. snake_case) are often preferred over periods or spaces. It is good practice to keep track of where you got files. The commands used to get these books are stored in the `README.md`. The `README.md`  file is written in [Markdown](https://en.wikipedia.org/wiki/Markdown). (To learn more about Markdown syntax, read this excellent [Markdown guide](https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet).)
+
+=== "Input"
+
+	```
+	head README.md
+	```
+
+=== "Expected Output"
+
+	```
+	# Books
+	
+	These books were downloaded from [Project 	Gutenberg](https://www.gutenberg.org/ebooks/) using the following commands. 
+	
+	
+	curl https://www.gutenberg.org/files/98/98-0.txt -o book.txt
+	curl https://www.gutenberg.org/files/98/98-0.txt -o A-tale-of-two-cities.txt
+	curl https://www.gutenberg.org/files/11/11-0.txt -o Alice_in_wonderland.txt
+	curl https://www.gutenberg.org/files/16/16-0.txt -o PeterPan.txt
+	curl https://www.gutenberg.org/files/55/55-0.txt -o WizardOfOz.txt
+	```
+
[…]
