### 2022-01-20 84f73f4
redirection and for loops
