+++ docs/General-Tools/UNIX/unix_6.md
@@ -1,0 +1,202 @@
+# Redirection and for loops
+
+If you did the last challenge, you saw that the **images/** directory contains a file called **MiSeq-readcount-Mothur.png**. This image is a screenshot from the [Mauther software tutorial](https://mothur.org/wiki/miseq_sop/) showing the count or number of reads for each sample. 
+
+![Mothur read count](./images/MiSeq-readcount-Mothur.png)
+
+
+To see if our data matches theirs, we can count the number of lines in the .fastq files with the UNIX command `wc`. This will print by default the number of characters, words, and lines in a file. We can ask for just the number of lines with the `-l`  option. 
+
+
+=== "Input"
+	
+	```
+	wc -l *fastq
+	```
+
+
+=== "Expected Output"
+
+
+	```
+	31172 F3D0_S188_L001_R1_001.fastq
+    31172 F3D0_S188_L001_R2_001.fastq
+    23832 F3D141_S207_L001_R1_001.fastq
+    23832 F3D141_S207_L001_R2_001.fastq
+    12732 F3D142_S208_L001_R1_001.fastq
+    12732 F3D142_S208_L001_R2_001.fastq
+    12712 F3D143_S209_L001_R1_001.fastq
+    12712 F3D143_S209_L001_R2_001.fastq
+    19308 F3D144_S210_L001_R1_001.fastq
+    19308 F3D144_S210_L001_R2_001.fastq
+    29508 F3D145_S211_L001_R1_001.fastq
+    29508 F3D145_S211_L001_R2_001.fastq
+    20084 F3D146_S212_L001_R1_001.fastq
+    20084 F3D146_S212_L001_R2_001.fastq
+    68280 F3D147_S213_L001_R1_001.fastq
+    68280 F3D147_S213_L001_R2_001.fastq
+    49620 F3D148_S214_L001_R1_001.fastq
+    49620 F3D148_S214_L001_R2_001.fastq
+    52332 F3D149_S215_L001_R1_001.fastq
+    52332 F3D149_S215_L001_R2_001.fastq
+    22036 F3D150_S216_L001_R1_001.fastq
+    22036 F3D150_S216_L001_R2_001.fastq
+    23476 F3D1_S189_L001_R1_001.fastq
+    23476 F3D1_S189_L001_R2_001.fastq
+    78480 F3D2_S190_L001_R1_001.fastq
+    78480 F3D2_S190_L001_R2_001.fastq
+    27032 F3D3_S191_L001_R1_001.fastq
+    27032 F3D3_S191_L001_R2_001.fastq
+    17792 F3D5_S193_L001_R1_001.fastq
+    17792 F3D5_S193_L001_R2_001.fastq
+    31956 F3D6_S194_L001_R1_001.fastq
+    31956 F3D6_S194_L001_R2_001.fastq
+    20516 F3D7_S195_L001_R1_001.fastq
+    20516 F3D7_S195_L001_R2_001.fastq
+    21176 F3D8_S196_L001_R1_001.fastq
+    21176 F3D8_S196_L001_R2_001.fastq
+    28280 F3D9_S197_L001_R1_001.fastq
+    28280 F3D9_S197_L001_R2_001.fastq
+    19116 Mock_S280_L001_R1_001.fastq
+    19116 Mock_S280_L001_R2_001.fastq
+  	1218880 total
+	```
+
+However, this number is too large. In fact, it is 4 times larger than the number of reads. To capture just the number of reads, we can first use `grep` followed `wc`.
+
+
+By default, many UNIX commands like `cat` send output to something called
+standard out, or "stdout". This is a catch-all phrase for "the basic
+place we send regular output." (There's also standard error, or "stderr",
+which is where errors are printed; and standard input, or "stdin", which
+is where input comes from.)
+
[…]
