### 2022-01-20 06e1ec3
summary or conclusing thoughts
