### 2022-01-21 c1451e0
add commands to front page
