### 2022-01-21 b33af3f
update training page description with link to lessons
