### 2022-01-21 bfe92d0
rm wc from find lesson
