+++ docs/General-Tools/UNIX/unix_5.md
@@ -1,0 +1,360 @@
+# Finding things
+
+A big part of data science is making sure what you _expect_ in a particular file is what you _have_ in that file. This is fairly easy when your files are small but is challenging when the files are much larger than your screen. 
+
+
+
+
+#### Books 
+
+Let's navigate to the books directory to explore this topic. 
+
+
+```
+cd ~/books
+pwd
+```
+
+
+Earlier, when we used `head` and `tail` to view the books, we read extraneous information about Project Gutenberg. Let's say you want to extract some information about each book such as the author or title? We can use the command `grep` to search for a pattern
+
+=== "Input"
+
+	```
+	grep Title *txt
+	```
+
+=== "Expected Output"
+
+	```
+	Alice_in_wonderland.txt:Title: Alice’s Adventures in Wonderland
+	A-tale-of-two-cities.txt:Title: A Tale of Two Cities
+	book.txt:Title: A Tale of Two Cities
+	PeterPan.txt:Title: Peter Pan
+	WizardOfOz.txt:Title: The Wonderful Wizard of Oz
+	```
+
+What are the Chapter titles of each book?
+
+=== "Input"
+
+	```
+	grep Chapter *txt
+	```
+
+=== "Expected Output"
+
+	You should see some text including the following:
+
+	```
+	PeterPan.txt: Chapter I. PETER BREAKS THROUGH
+	PeterPan.txt: Chapter II. THE SHADOW
+	...
+	WizardOfOz.txt: Chapter I. The Cyclone
+	WizardOfOz.txt: Chapter II. The Council with the Munchkins
+
+	```
+
+Why do we only see Chapter titles for 2 books? Because the other Chapter titles are written in all caps. Let's modify grep to match Chapter 1, chapter 1, and CHAPTER 1. We use the `-i` option to "ignore case" and the `-w` option to match the word. 
+
+=== "Input"
+
+	```
+	grep -i -w "chapter i" *txt
+	```
+
+=== "Expected Output"
+
+
+	This returns the first chapter for each book. Shown below is the  first chapter for each book. This pattern occurs once in the table of contents and once in the main text. A Tale of Two Cities has 3 Chapter 1s. 
+
+
+	```
+	Alice_in_wonderland.txt: CHAPTER I.     Down the Rabbit-Hole
+	Alice_in_wonderland.txt:CHAPTER I.
+	A-tale-of-two-cities.txt:     CHAPTER I      The Period
+	A-tale-of-two-cities.txt:     CHAPTER I      Five Years Later
+	A-tale-of-two-cities.txt:     CHAPTER I      In Secret
+	A-tale-of-two-cities.txt:CHAPTER I.
+	A-tale-of-two-cities.txt:CHAPTER I.
+	A-tale-of-two-cities.txt:CHAPTER I.
+	book.txt:     CHAPTER I      The Period
+	book.txt:     CHAPTER I      Five Years Later
+	book.txt:     CHAPTER I      In Secret
+	book.txt:CHAPTER I.
+	book.txt:CHAPTER I.
+	book.txt:CHAPTER I.
+	PeterPan.txt: Chapter I. PETER BREAKS THROUGH
[…]
