### 2022-01-22 461db28
update home training page
