### 2022-01-25 #132 @willingc
@betatim @choldgraf @sgibson91 Do any of you have time to do a review of this paper? Requires some knowledge of AWS. I would really appreciate it. Thanks.

@abhijna I'm back now so this should get back on track. Sorry for the delay.
