### 2022-01-25 0492df9
Create rna_seq_9.md
