+++ docs/Bioinformatic-Analyses/RNAseq-on-Cavatica/rna_seq_9.md
@@ -1,0 +1,176 @@
+---
+layout: page
+title: Incorporating GTEx Data in Kids First Analyses
+---
+
+<div class="banner"><span class="banner-text">Lesson in Development</span></div>
+
+Analysis using Data Cruncher
+==========================
+
+So far we have explored running DGE analysis using a public app based on DESeq2. In the second approach, we will set up an interactive analysis on an instance running the RStudio computational environment. We will run a DGE workflow using an analysis script to generate reports and plots.
+
+!!! note "DGE Tools"
+
+    While there are other established tools to perform DGE analysis including [DESeq2](https://bioconductor.org/packages/release/bioc/html/DESeq2.html), [EdgeR](https://bioconductor.org/packages/release/bioc/html/edgeR.html) and [Limma-Voom](https://genomebiology.biomedcentral.com/articles/10.1186/gb-2014-15-2-r29), we will be using DESeq2 in our script to allow you to compare the output between the two approaches.
+
+
+## Step 1: Starting Data Cruncher
+
+* Click the <span class="highlight_txt">Interactive Analysis</span> tab located on the right hand corner below your account settings menu.
+* Select <span class="highlight_txt">Open</span> in the Data Cruncher panel.
+* Click on the <span class="highlight_txt">Create your first analysis</span> which appears the first time your are setting up.
+* In the popup box, select <span class="highlight_txt">RStudio</span> for Environment. Provide an analysis name in the box. Here, **Cancer_DGE** was used to title the analysis. Click <span class="highlight_txt">Next</span> when done.
+* In the `Compute requirements` tab, we will use the default instance type (c5.2xlarge, $0.49/hr). We increase the `Suspend time`, which is the period of inactivity after which the instance is stopped automatically and the analysis is saved, from 30 to 60 minutes.
+* Click <span class="highlight_txt">Start the analysis</span>. This prompts the initialization of the analysis which involves set up of the instance and preparation of the analysis environment.
+
+<iframe id="kaltura_player" src="https://cdnapisec.kaltura.com/p/1770401/sp/177040100/embedIframeJs/uiconf_id/29032722/partner_id/1770401?iframeembed=true&playerId=kaltura_player&entry_id=1_4hc4cd8i&flashvars[mediaProtocol]=rtmp&amp;flashvars[streamerType]=rtmp&amp;flashvars[streamerUrl]=rtmp://www.kaltura.com:1935&amp;flashvars[rtmpFlavors]=1&amp;flashvars[localizationCode]=en&amp;flashvars[leadWithHTML5]=true&amp;flashvars[sideBarContainer.plugin]=true&amp;flashvars[sideBarContainer.position]=left&amp;flashvars[sideBarContainer.clickToClose]=true&amp;flashvars[chapters.plugin]=true&amp;flashvars[chapters.layout]=vertical&amp;flashvars[chapters.thumbnailRotator]=false&amp;flashvars[streamSelector.plugin]=true&amp;flashvars[EmbedPlayer.SpinnerTarget]=videoHolder&amp;flashvars[dualScreen.plugin]=true&amp;flashvars[mediaProxy.mediaPlayTo]=44&amp;flashvars[Kaltura.addCrossoriginToIframe]=true&amp;&wid=1_6ih50uzu" width="608" height="402" allowfullscreen webkitallowfullscreen mozAllowFullScreen allow="autoplay *; fullscreen *; encrypted-media *" sandbox="allow-forms allow-same-origin allow-scripts allow-top-navigation allow-pointer-lock allow-popups allow-modals allow-orientation-lock allow-popups-to-escape-sandbox allow-presentation allow-top-navigation-by-user-activation" frameborder="0" title="Kaltura Player"></iframe>
+
+!!! info "Instance Types"
+
[…]
