### 2022-01-25 768e4df
Update rna_seq_9.md
