### 2022-01-25 e97d945
Update rna_seq_9.md
