+++ hackmd/AWS.md
@@ -1,335 +1,515 @@
-# A Hands-on Introduction to AWS
+# A Hands-on Introduction to AWS for Cloud Computing
 
-**When:**  
+**When:**   January 26, 2022, from 10 am PST - 12 pm PST 
 
-**Instructors:**  
+**Instructors:** Dr. Rayna Harris
 
-**Moderator:**  
+**Helpers:**  Dr. Amanda Charbonneau, Jessica Lumian, Jeremy Walter 
 
-**Helpers:**  
-
-## Description
-
-This 2 hour hands-on tutorial will introduce you to creating a computer "in the cloud" and logging into it, via Amazon Web Services. We'll create a small general-purpose Linux computer, connect to it, and run a small job while discussing the concepts and technologies involved.
+Your instructors are part of the training and engagement team for the [NIH Common Fund Data Ecosystem](https://nih-cfde.org/), a project supported by the NIH to increase data reuse and cloud computing for biomedical research.
 
 
-> :smiley_cat: While we wait to get started
-> 1. :heavy_check_mark: Have you checked out the [pre-workshop resources page](https://github.com/nih-cfde/training-and-engagement/wiki/Resources-for-Workshop-Attendees)?
-> 2. :pencil: Please fill out our pre-workshop survey if you have not already done so! [Please click this link](https://forms.gle/hCtbXrDyH9omAYyB6) 
-> 3. If you are on a windows computer, make sure you have Mobaxterm. [Check out the install guide!](https://hackmd.io/YivVGhznRwqVO3fxdGpkOQ?view)
+#### Description
+
+This 2-hour hands-on tutorial will introduce you to creating a computer "in the cloud" and logging into it, via Amazon Web Services. We will launch a small general-purpose Linux instance, connect to it, and run a small job while discussing the concepts and technologies involved.
+
+#### Overview
+
+[TOC]
+
+#### Before we start
+
+:pencil: Please fill out our [pre-workshop survey](https://forms.gle/hCtbXrDyH9omAYyB6) if you have not already done so! 
+
+:heavy_check_mark: Windows users should install Mobaxterm. Read our [quick installation guide.](https://hackmd.io/YivVGhznRwqVO3fxdGpkOQ?view)
+
+#### Questions?
+
+If you have questions,  
+- Type them in the group chat
+- Direct message the moderator 
+- Unmute and ask them outloud
+
+We're going to use the raised hand :raised_hand: reaction in zoom to make sure people are on board during the hands-on activities.
+
+## 1. Terminology and Sign-On
+
+Cloud computing is the on-demand use of data storage and compute power without direct active management by the user. Amazon Web Services (AWS) is one of the most broadly adopted cloud platforms.
 
 
-## Hello!
+Amazon's Elastic Compute Cloud (**EC2**) is a web service that provides secure, resizable compute capacity in the cloud. Amazon's Simple Storage Service (**S3**) is widely used for storing and sharing data. An **instance** is a virtual machine that runs in the cloud. An **image** is a resource that stores the configuration files, metadata, permissions, and data to create, maintain, backup, or recover an instance.
 
-Your instrutors are part of the training and engagement team for the [NIH Common Fund Data Ecosystem](https://nih-cfde.org/), a project supported by the NIH to increase data reuse and cloud computing for biomedical research.
+Some advantages of using AWS include:
 
-
-> 👍 **Have you heard of the NIH Common Fund Data Ecosystem?**
-> Put up a :heavy_check_mark: for yes and a :negative_squared_cross_mark: for no!
-
-We have the following goals for this workshop:
-
-* Help you think about if and how to use cloud computers for your work!
-* Gather questions, feedback and refine the tutorial materials!
-
[…]
