### 2022-01-26 5b7297f
Update rna_seq_9.md
