### 2022-01-27 2506e5f
Update rna_seq_9.md
