### 2022-01-27 78b1e7d
Update rna_seq_9.md
