### 2022-01-27 a240613
Update rna_seq_9.md
