### 2022-01-27 981bb30
Update rna_seq_9.md
