### 2022-01-27 298520a
Update rna_seq_9.md
