### 2022-01-27 38673dd
minor updates to rna_seq_9.md
two minor updates
Section 2
Change from: 
- Once the PFB is created, download it to your local machine. This document [...]

to:
- Once the PFB is created, download it to your local machine. This downloaded PFB file [...]

Section 3
Change from:
The commands for importing the GTEx files from AnVIL into CAVATICA are written into the `interop.py` script.

to:
The commands for importing the GTEx files from AnVIL into CAVATICA are written into the `interop.py` script included in the public project you copied.
