+++ hackmd/AWS.md
@@ -1,13 +1,12 @@
 # A Hands-on Introduction to AWS for Cloud Computing
 
-**When:**    
+**When:**   January 26, 2022, from 10 am PST - 12 pm PST 
 
-**Instructors:**  
+**Instructors:** Dr. Rayna Harris
 
-**Helpers:**  
+**Helpers:** Dr. Amanda Charbonneau  Jessica Lumian and Jeremy Walter 
 
 Your instructors are part of the training and engagement team for the [NIH Common Fund Data Ecosystem](https://nih-cfde.org/), a project supported by the NIH to increase data reuse and cloud computing for biomedical research.
-
 
 #### Description
 
@@ -36,17 +35,21 @@ We're going to use the raised hand :raised_hand: reaction in zoom to make sure p
 
 Cloud computing is the on-demand use of data storage and compute power without direct active management by the user. Amazon Web Services (AWS) is one of the most broadly adopted cloud platforms.
 
-
-Amazon's Elastic Compute Cloud (**EC2**) is a web service that provides secure, resizable compute capacity in the cloud. Amazon's Simple Storage Service (**S3**) is widely used for storing and sharing data. An **instance** is a virtual machine that runs in the cloud. An **image** is a resource that stores the configuration files, metadata, permissions, and data to create, maintain, backup, or recover an instance.
-
 Some advantages of using AWS include:
 
 - Easy sign-on 
 - Simple billing
 - Stable services
+- Customizable images
 - Customer support
 - Online resources
 
+
+![](https://uploads-ssl.webflow.com/5e1f17bab0dc6527c1ecc801/5e55f0ab6725fd082d2ea435_amazon-hosting.jpeg)
+
+Amazon's Elastic Compute Cloud (**EC2**) is a web service that provides secure, resizable compute capacity in the cloud. Amazon's Simple Storage Service (**S3**) is widely used for storing and sharing data. 
+
+An **instance** is a virtual machine that runs in the cloud. An **image** (or AMI for Amazon Machine Image) is a template that contains the software configuration (including operating system and applications) required to launch your instance. You can select an image provided by the AWS Marketplace, the AWS community, or you can select one of your own images. When you launch an instance, you specify the type of image to use. 
 
 Today, everything you do will be paid for by us. Your free login credentials will work for the next 24 hours. In the future, if you create an AWS account, you will have to add a credit card for billing. We'd be happy to answer questions about how to pay for AWS.
 
@@ -54,7 +57,21 @@ Log in to your account by going to this web address:  https://cfde-training-work
 
 ![](https://hackmd.io/_uploads/SJfyT66pt.png)
 
-Find your first name in the table below and log in with that as your IAM user name and the password given by your instructor.
+Find your first name in the table below and log in with that as your IAM user name and the password `cfde2022!!` 
+
+
+|   IAM |   IAM  |   IAM  |   IAM  |   Password  |
+|---|---|---|---|---|
+|   Abdullahi  |   Jasleen  |   Minoo  |   Somi  |   `cfde2022!!`  |
+|   Brett  |   Jenna  |   Nathan  |   Sophia  |   `cfde2022!!`  |
+|   Connor  |   Jesus  |   Nicole  |   Stefan  |   `cfde2022!!`   |
+|   Daulet  |   Jiefei  |   Nina  |   Stephen  |   `cfde2022!!`   |
+|   David  |   Ketaki  |   Owen  |   Triveni  |   `cfde2022!!`   |
+|   Francisco  |   Kirtan   |   Paul  |   Uma  |   `cfde2022!!`   |
[…]
