+++ docs/Common-Fund-Tools/IDG.md
@@ -1,0 +1,58 @@
+# Illuminating the Druggable Genome (IDG) Tutorials
+
+
+![](RackMultipart20220128-4-1vi3wev_html_dd4632e6b684fa0a.png)
+
+The NIH funded the Illuminating the Druggable Genome (IDG) program to improve our understanding of the properties and functions of proteins that are currently under-annotated within the three most commonly drug-targeted protein families: G-protein coupled receptors, ion channels, and protein kinases. Since 2014, the IDG Knowledge Management Center (IDG-KMC) has generated several open-access datasets and resources that jointly serve as a highly translational machine-learning-ready knowledgebases focused on human protein-coding genes and their products. The goal of the IDG-KMC is to develop comprehensive integrated knowledge for the druggable genome to illuminate the uncharacterized or poorly annotated portion of the druggable genome. The tools derived from the IDG-KMC provide either user-friendly visualizations or ways to impute knowledge about potential targets using machine learning strategies.
+
+For step-by-step descriptions on how to use each web-based tool to accelerate illumination in under-studied proteins check out our [Current Protocols doi: 10.1002/cpz1.355](https://currentprotocols.onlinelibrary.wiley.com/doi/10.1002/cpz1.355).
+
+Visit https://druggablegenome.net/ for more information on IDG Consortium.
+
+## IDG Interacting with the Pharos user interface
+
+[https://youtu.be/1FKp0gskGhU](https://youtu.be/1FKp0gskGhU) Presenter: Tim Sheils, NCATS (IDG KMC)
+
+Demo outcome: Participants will learn how to browse through Pharos, upload and save lists, analyze lists, and get predicted ligand interactions. IDG Grant: KMC-UNM U24CA224370 and publication: https://doi.org/10.1093/nar/gkaa993
+
+Link to Digital tool: https://pharos.nih.gov/
+
+[https://youtu.be/EVir\_va8xaY](https://youtu.be/EVir_va8xaY) (Research Database) with GraphQL API methodology allowing for more fine-tuned queries. Presenters: Keith Kelleher, NCATS (IDG KMC) Demo Outcome: Participants will learn how to query the Pharos API using GraphQL. IDG Grant: KMC-UNM U24CA224370 Link to Digital tool:[https://pharos.nih.gov/api](https://www.youtube.com/redirect?event=video_description&amp;redir_token=QUFFLUhqbmhwcnZKdDdrdjd3VVJhdFZhTlFwV3BCT1cyQXxBQ3Jtc0tuNXRDbkRxbTJHNC00MU0wTTEwVE1uMVhkRE1JNnMtN04xU0lwX2V6ZHBuUFBMdVY5akJfMkZZUXJKRjFueTBiRGZnRmROQzYtc1g3bWk5RVhvaW80ai1MWnp1QWhscTROVkhYMFpodWJRTzhINThEbw&amp;q=https%3A%2F%2Fpharos.nih.gov%2Fapi)
+
+## The Harmonizome resource
+
[…]
