+++ docs/Common-Fund-Tools/IDG.md
@@ -1,58 +1,130 @@
 # Illuminating the Druggable Genome (IDG) Tutorials
 
+![](../images/IDG_Logo.png)
 
-![](RackMultipart20220128-4-1vi3wev_html_dd4632e6b684fa0a.png)
+The NIH funded the [Illuminating the Druggable Genome (IDG)](https://druggablegenome.net/) program to improve our understanding of the properties and functions of proteins that are currently under-annotated within the three most commonly drug-targeted protein families: G-protein coupled receptors, ion channels, and protein kinases. 
 
-The NIH funded the Illuminating the Druggable Genome (IDG) program to improve our understanding of the properties and functions of proteins that are currently under-annotated within the three most commonly drug-targeted protein families: G-protein coupled receptors, ion channels, and protein kinases. Since 2014, the IDG Knowledge Management Center (IDG-KMC) has generated several open-access datasets and resources that jointly serve as a highly translational machine-learning-ready knowledgebases focused on human protein-coding genes and their products. The goal of the IDG-KMC is to develop comprehensive integrated knowledge for the druggable genome to illuminate the uncharacterized or poorly annotated portion of the druggable genome. The tools derived from the IDG-KMC provide either user-friendly visualizations or ways to impute knowledge about potential targets using machine learning strategies.
+Since 2014, the IDG Knowledge Management Center (IDG-KMC) has generated several open-access datasets and resources that jointly serve as a highly translational machine-learning-ready knowledgebases focused on human protein-coding genes and their products. The goal of the IDG-KMC is to develop comprehensive integrated knowledge for the druggable genome to illuminate the uncharacterized or poorly annotated portion of the druggable genome. 
 
-For step-by-step descriptions on how to use each web-based tool to accelerate illumination in under-studied proteins check out our [Current Protocols doi: 10.1002/cpz1.355](https://currentprotocols.onlinelibrary.wiley.com/doi/10.1002/cpz1.355).
+The tools derived from the IDG-KMC provide either user-friendly visualizations or ways to impute knowledge about potential targets using machine learning strategies. For step-by-step descriptions on how to use each web-based tool to accelerate illumination in under-studied proteins check out our [Current Protocols doi: 10.1002/cpz1.355](https://currentprotocols.onlinelibrary.wiley.com/doi/10.1002/cpz1.355).
 
-Visit https://druggablegenome.net/ for more information on IDG Consortium.
+Here, we provide links to and descriptions of some recent video tutorials and presentations on IDG tools and resources. You can also subscribe to the [IDG YouTube Channel](https://www.youtube.com/c/DruggableGenomeIDG/videos) to stay up-to-date on new topics.  
 
-## IDG Interacting with the Pharos user interface
+## Interacting with the Pharos User Interface
 
[…]
