+++ docs/Common-Fund-Tools/IDG.md
@@ -17,8 +17,8 @@ Here, we provide links to and descriptions of some recent video tutorials and pr
 * Presenter: Tim Sheils, NCATS (IDG KMC)
 * Demo outcome: Participants will learn how to browse through [Pharos](https://pharos.nih.gov/), upload and save lists, analyze lists, and get predicted ligand interactions. 
 * IDG Grant: KMC-UNM U24CA224370 and publication: https://doi.org/10.1093/nar/gkaa993
-* Link to digital tool: https://pharos.nih.gov/
-* Link to video tuorial: https://youtu.be/1FKp0gskGhU 
+* Link to digital tool: <https://pharos.nih.gov/>
+* Link to video tuorial: <https://youtu.be/1FKp0gskGhU> 
 
 ## Pharos GraphQL API
 
@@ -27,8 +27,8 @@ Here, we provide links to and descriptions of some recent video tutorials and pr
 * Presenters: Keith Kelleher, NCATS (IDG KMC) 
 * Demo Outcome: Participants will learn how to query the [Pharos API](https://pharos.nih.gov/api) using GraphQL. 
 * IDG Grant: KMC-UNM U24CA224370 
-* Link to digital tool: https://pharos.nih.gov/api
-* Link to video tuorial: https://youtu.be/EVir_va8xaY
+* Link to digital tool: <https://pharos.nih.gov/api>
+* Link to video tuorial: <https://youtu.be/EVir_va8xaY>
 
 ## The Harmonizome Resource
 
@@ -37,10 +37,10 @@ Here, we provide links to and descriptions of some recent video tutorials and pr
 * Presenter: Avi Ma'ayan
 * Description: [Harmonizome](https://maayanlab.cloud/Harmonizome) is a collection of processed datasets gathered to serve and mine knowledge about genes and proteins from over 70 major online resources. We extracted, abstracted and organized data into ∼72 million functional associations between genes/proteins and their attributes. 
 * Demo Outcome: Understanding the concepts behind constructing the Harmonizome. 
-* Publication: https://doi.org/10.1093/database/baw100
+* Publication: <https://doi.org/10.1093/database/baw100>
 * IDG Grant: KMC-ISMMS U24CA224260
-* Link to digital tool: https://maayanlab.cloud/Harmonizome
-* Link to video tuorial: https://www.youtube.com/watch?v=yGkIQjeWh9U&t=26s
+* Link to digital tool: <https://maayanlab.cloud/Harmonizome>
+* Link to video tuorial: <https://www.youtube.com/watch?v=yGkIQjeWh9U&t=26s>
 
 
 ## The ARCHS4 Fesource
@@ -60,9 +60,9 @@ Here, we provide links to and descriptions of some recent video tutorials and pr
 
 * Presenter: Alexander Lachmann
 * Description: In this e-IDG lecture Dr. Alexander Lachmann, an Assistant Professor from the Ma&#39;ayan Lab at ISMMS, describes three projects: [PrismExp](https://appyters.maayanlab.cloud/#/PrismEXP), blitzGSEA, and bridgeGSEA. These are three related methods to predict gene function and incorporate such predictions within gene set enrichment analyses. 
-* Publication: https://www.biorxiv.org/content/10.11
-* Link to digital tool: https://appyters.maayanlab.cloud/#/PrismEXP
-* Link to video: https://youtu.be/6gDzA2dfOrI
+* Publication: <https://www.biorxiv.org/content/10.11>
+* Link to digital tool: <https://appyters.maayanlab.cloud/#/PrismEXP>
+* Link to video: <https://youtu.be/6gDzA2dfOrI>
 
 ## Using Geneshot to illuminate knowledge about understudied targets
 
[…]
