+++ docs/Common-Fund-Tools/IDG.md
@@ -4,9 +4,10 @@
 
 The NIH funded the [Illuminating the Druggable Genome (IDG)](https://druggablegenome.net/) program to improve our understanding of the properties and functions of proteins that are currently under-annotated within the three most commonly drug-targeted protein families: G-protein coupled receptors, ion channels, and protein kinases. 
 
-Since 2014, the IDG Knowledge Management Center (IDG-KMC) has generated several open-access datasets and resources that jointly serve as a highly translational machine-learning-ready knowledgebases focused on human protein-coding genes and their products. The goal of the IDG-KMC is to develop comprehensive integrated knowledge for the druggable genome to illuminate the uncharacterized or poorly annotated portion of the druggable genome. 
+Since 2014, the IDG Knowledge Management Center (IDG-KMC) has generated several open-access datasets and resources that jointly serve as a highly translational machine-learning-ready knowledgebase focused on human protein-coding genes and their products. The goal of the IDG-KMC is to develop comprehensive integrated knowledge for the druggable genome to illuminate the uncharacterized or poorly annotated portion of the druggable genome. 
 
 The tools derived from the IDG-KMC provide either user-friendly visualizations or ways to impute knowledge about potential targets using machine learning strategies. For step-by-step descriptions on how to use each web-based tool to accelerate illumination in under-studied proteins check out our [Current Protocols doi: 10.1002/cpz1.355](https://currentprotocols.onlinelibrary.wiley.com/doi/10.1002/cpz1.355).
+
 
 Here, we provide links to and descriptions of some recent video tutorials and presentations on IDG tools and resources. You can also subscribe to the [IDG YouTube Channel](https://www.youtube.com/c/DruggableGenomeIDG/videos) to stay up-to-date on new topics.  
 
@@ -18,7 +19,7 @@ Here, we provide links to and descriptions of some recent video tutorials and pr
 * Demo outcome: Participants will learn how to browse through [Pharos](https://pharos.nih.gov/), upload and save lists, analyze lists, and get predicted ligand interactions. 
 * IDG Grant: KMC-UNM U24CA224370 and publication: https://doi.org/10.1093/nar/gkaa993
 * Link to digital tool: <https://pharos.nih.gov/>
-* Link to video tuorial: <https://youtu.be/1FKp0gskGhU> 
+* Link to video tutorial: <https://youtu.be/1FKp0gskGhU> 
 
 ## Pharos GraphQL API
 
@@ -28,7 +29,7 @@ Here, we provide links to and descriptions of some recent video tutorials and pr
 * Demo Outcome: Participants will learn how to query the [Pharos API](https://pharos.nih.gov/api) using GraphQL. 
 * IDG Grant: KMC-UNM U24CA224370 
 * Link to digital tool: <https://pharos.nih.gov/api>
-* Link to video tuorial: <https://youtu.be/EVir_va8xaY>
+* Link to video tutorial: <https://youtu.be/EVir_va8xaY>
 
 ## The Harmonizome Resource
 
@@ -40,15 +41,15 @@ Here, we provide links to and descriptions of some recent video tutorials and pr
 * Publication: <https://doi.org/10.1093/database/baw100>
 * IDG Grant: KMC-ISMMS U24CA224260
 * Link to digital tool: <https://maayanlab.cloud/Harmonizome>
-* Link to video tuorial: <https://www.youtube.com/watch?v=yGkIQjeWh9U&t=26s>
+* Link to video tutorial: <https://www.youtube.com/watch?v=yGkIQjeWh9U&t=26s>
 
 
-## The ARCHS4 Fesource
+## The ARCHS4 Resource
 
[…]
