+++ docs/Common-Fund-Tools/index.md
@@ -8,10 +8,8 @@ Common Fund Tools
 Tutorials in this section provide lessons for finding data sets across the NIH Common Fund.
 
 - [Kids First Tutorials](Kids-First/index.md)
-
-> Lessons in Development:
-
-> - [CFDE Portal Tutorials](CFDE-Portal/index.md)
+- [Illuminating the Druggable Genome (IDG) Tutorials](IDG.md)
+- [CFDE Portal Tutorials](CFDE-Portal/index.md) (in development)
 
 You may also be interested in this lesson, which uses Common Fund program data:
 
