+++ docs/Common-Fund-Tools/index.md
@@ -7,9 +7,10 @@ Common Fund Tools
 
 Tutorials in this section provide lessons for finding data sets across the NIH Common Fund.
 
+- [CFDE Portal Tutorials](CFDE-Portal/index.md) (in development)
 - [Kids First Tutorials](Kids-First/index.md)
 - [Illuminating the Druggable Genome (IDG) Tutorials](IDG.md)
-- [CFDE Portal Tutorials](CFDE-Portal/index.md) (in development)
+
 
 You may also be interested in this lesson, which uses Common Fund program data:
 
