### 2022-02-12 4d4a780
Update rna_seq_9.md
added requested change, Click the "Explore Data" button to use AnVIL Data Explorer
