### 2022-02-12 ee3c98a
Update rna_seq_9.md
line 23 changed `Export All to PFB` to `Export to PFB`
