### 2022-02-12 88f7eb2
Update rna_seq_9.md
corrected image reference on line 28 from a Cavatica login screen on rna-seq-9-06.png to an AnVIL Create API Key
