### 2022-02-12 4e30cd6
Update rna_seq_9.md
corrected Step 2 screenshots -- they were offset by one, starting with "Navigate to Cavatica" which needed to be changed from rna-seq-9-07.png to rna-seq-9-06.png up until "clicking to enter Data Cruncher" from rna-seq-9-13.png to rna-seq-9-12.png
