### 2022-02-12 6805863
Update rna_seq_9.md
corrected path for NIH and gen3 screenshot images
