### 2022-02-12 46eee19
Update rna_seq_9.md
corrected missing .png file extension for rna-seq-9-06-02-gen3-authorize.png
