### 2022-02-12 4d27bf5
Update rna_seq_9.md
replaced old rna-seq-9-14.png with rna-seq-9-14-AnVIL_transfer_ipynb.png
