### 2022-02-12 56332d3
Update rna_seq_9.md
remove extra ] after Portal
