+++ docs/Bioinformatic-Analyses/RNAseq-on-Cavatica/rna_seq_9.md
@@ -17,7 +17,7 @@ The previous example provides a walkthrough for identifying RNA-Seq datasets fro
 
 ## Step 1: Identify GTEx Files on AnVIL and Export a PFB
 - Navigate to [the AnVIL Portal](https://gen3.theanvil.io/login) and login using NIH or Google credentials.
-    - <img src="./rna-seq-images/rna-seq-9-01.png" width="400"/>
+    - ![portal](./rna-seq-images/rna-seq-9-01.png){: align=left width=50%}
 - Click the "Explore Data" button to use the [AnVIL Data Explorer](https://gen3.theanvil.io/explorer), identify the GTEx files for your analysis. For more information on using AnVIL Gen3, see [their support documentation here](https://anvilproject.org/learn/introduction/getting-started-with-gen3).
     - <img src="./rna-seq-images/rna-seq-9-02.png" width="400"/>
 - Once you have identified the subset of files for your analysis, return to the top of the screen and choose `Export to PFB`. This will create a manifest of files on the AnVIL Portal. This process may take several minutes, depending upon the number of files for export.
