+++ docs/Bioinformatic-Analyses/RNAseq-on-Cavatica/rna_seq_9.md
@@ -18,15 +18,15 @@ The previous example provides a walkthrough for identifying RNA-Seq datasets fro
 ## Step 1: Identify GTEx Files on AnVIL and Export a PFB
 
 - Navigate to [the AnVIL Portal](https://gen3.theanvil.io/login) and login using NIH or Google credentials.
-    ![one](./rna-seq-images/rna-seq-9-01.png){: align=right width=95%}
+    ![one](./rna-seq-images/rna-seq-9-01.png)
 - Click the "Explore Data" button to use the [AnVIL Data Explorer](https://gen3.theanvil.io/explorer), identify the GTEx files for your analysis. For more information on using AnVIL Gen3, see [their support documentation here](https://anvilproject.org/learn/introduction/getting-started-with-gen3).
-    ![two docs](./rna-seq-images/rna-seq-9-02.png){: align=right width=95%}
+    ![two docs](./rna-seq-images/rna-seq-9-02.png)
 - Once you have identified the subset of files for your analysis, return to the top of the screen and choose `Export to PFB`. This will create a manifest of files on the AnVIL Portal. This process may take several minutes, depending upon the number of files for export.
-    ![three](./rna-seq-images/rna-seq-9-03.png){: align=right width=95%}
+    ![three](./rna-seq-images/rna-seq-9-03.png)
 - Once the PFB is created, download it to your local machine. This downloaded PFB file contains the information CAVATICA needs to pull these files into your CAVATICA project.
-    ![four](./rna-seq-images/rna-seq-9-04.png){: align=right width=95%}
+    ![four](./rna-seq-images/rna-seq-9-04.png)
 - Before navigating away from the AnVIL Portal, you also must generate an API key. This will be used to authenticate your access to the files in the PFB and allow the transfer of GTEx data into CAVATICA. Select your username in the top right of the AnVIL Portal. Select `Create API Key` in the top left of the page that opens. After a key is generated, download this key by choosing `Dowload json`. A file will save to your local machine.
-    ![five](./rna-seq-images/rna-seq-9-05.png){: align=right width=95%}
+    ![five](./rna-seq-images/rna-seq-9-05.png)
 
 - This concludes the necessary steps in the AnVIL Portal. We will now move to CAVATICA.
 
@@ -35,29 +35,29 @@ The previous example provides a walkthrough for identifying RNA-Seq datasets fro
 The commands necessary to import the chosen GTEx files from the AnVIL Portal have been saved in a Jupyter Notebook that is available on the CAVATICA platform. We will copy this notebook to a project we can work in and then begin an interactive session in the Data Cruncher to run the transfer.
 
 - Navigate to [CAVATICA](https://cavatica.sbgenomics.com/) and login using NIH or CAVATICA credentials.
-    ![six](./rna-seq-images/rna-seq-9-06.png){: align=right width=95%}
+    ![six](./rna-seq-images/rna-seq-9-06.png)
     
 - Note: if you are logging into CAVATICA for the first time you will presented with an NIH consent screen followed by a Gen3 authorization screen.
-    ![consent](./rna-seq-images/rna-seq-9-06-01-nih-consent.png){: align=right width=95%}
-    ![auth](./rna-seq-images/rna-seq-9-06-02-gen3-authorize.png){: align=right width=95%}
+    ![consent](./rna-seq-images/rna-seq-9-06-01-nih-consent.png)
+    ![auth](./rna-seq-images/rna-seq-9-06-02-gen3-authorize.png)
 
 - Within CAVATICA, choose [`Public Projects`](https://cavatica.sbgenomics.com/public/projects) from the menu bar at the top, then select [`Data Interoperability`](https://cavatica.sbgenomics.com/u/sevenbridges/data-interoperability).
[…]
