+++ hackmd/README.md
@@ -1,15 +1,15 @@
 # README.md
 
-This directory is for syncing the HackMD versions of lessons with GitHub so that we can track changes and maintina a source file. 
+This directory is for syncing the HackMD versions of lessons with GitHub so that we can track changes and maintain a source file. 
 
 ## Workflow
 
 - From the master branch, open the lesson file of interest (e.g. UNIX.md, AWD.md) and click the edit button.
 - Make a change and save it to a new branch named after the specific workshop (e.g. jan-workshop, feb-workshop). Do NOT create a PR.
-- Go to [HackMd](https://hackmd.io/). Ckick "New note" then "Pull from GitHub". Select the GitHub repo and the lesson file. (You made need to authorize a new repo if this is your first time.) Accept all the changes. 
+- Go to [HackMd](https://hackmd.io/). Ckick "New note" then "Pull from GitHub". Select the GitHub repo and the lesson file. (You may need to authorize a new repo if this is your first time.) Accept all the changes. 
 - Modify the HackMD for the workshop. Make as many commits as you like. After the workshop, push the final version to GitHub.
 - In GitHub, edit the lesson on the workshop-specific branch to remove any dates or other workshop-specific details.  
-- Open a PR to merge any chagnes that were made to the lesson into the master branch. Merge the PR. 
+- Open a PR to merge any changes that were made to the lesson into the master branch. Merge the PR. 
 
 ## Lesson source files for HackMd
 
