### 2022-02-16 b434875
Create December-2021.md
