+++ docs/Release-Notes/December-2021.md
@@ -1,0 +1,20 @@
+---
+layout: page
+title: December 2021 Release
+---
+
+Latest Release
+=================
+
+**Updated December 7, 2021**
+
+<span style="font-size:24px;">Updates and Fixes
+
+- [Introduction to Amazon Web Services](../Cloud-Platforms/Introduction_to_Amazon_Web_Services/introtoaws1.md): Updated to reflect the changes in the AWS user experience interface
+- [Introduction to Conda](../General-Tools/Introduction-to-Conda/index.md): Added --from-history export as an additional option and included new vidlets for prominent steps
+- [RNAseq on Cavatica](../Bioinformatic-Analyses/RNAseq-on-Cavatica/rna_seq_1.md): Updated reference genome and added steps to resolve network settings
+
+<span style="font-size:24px;">Website Features
+
+- Workshops: Updated the past workshops list
+- Mailing list: Added mailing list email on home page and on footer
