### 2022-02-16 d6bb272
Delete ISSUE_TEMPLATE.md
