+++ docs/Release-Notes/June-2021.md
@@ -3,7 +3,7 @@ layout: page
 title: June 2021 Release
 ---
 
-Latest Release
+June 2021 Release
 =================
 
 **Updated June 25, 2021**
