+++ docs/Release-Notes/index.md
@@ -9,9 +9,9 @@ title: February 2022 Release
 
 ## New Tutorials
 
-- [Incorporating GTEx Data in Kids First Analyses](../Bioinformatic-Analyses/RNAseq-on-Cavatica/rna_seq_9.md). This lesson demonstrates how to identify GTEx Datasets from the AnVIL Portal and import them into the CAVATICA/Kids First environment. It builds on the existing RNA-Seq training documentation by providing directions for expanding analyses to include GTEx's data as a source of RNA-Seq controls to compare to Kids First tumor samples. It also demonstrates interoperability between two of CFDE's DCCs.
+- [Incorporating GTEx Data in Kids First Analyses](../Bioinformatic-Analyses/RNAseq-on-Cavatica/rna_seq_9.md): This lesson demonstrates how to identify GTEx Datasets from the AnVIL Portal and import them into the CAVATICA/Kids First environment. It builds on the existing RNA-Seq training documentation by providing directions for expanding analyses to include GTEx's data as a source of RNA-Seq controls to compare to Kids First tumor samples. It also demonstrates interoperability between two of CFDE's DCCs.
 
-- [IDG Tutorials](../Common-Fund-Tools/IDG.md). The tools derived from the Illuminating the Druggagle Genome (IDG) program provide either user-friendly visualizations or ways to impute knowledge about potential targets using machine learning strategies. Watch these videos for step-by-step descriptions on how to use each web-based tool to accelerate illumination in under-studied proteins.
+- [IDG Tutorials](../Common-Fund-Tools/IDG.md): The tools derived from the Illuminating the Druggagle Genome (IDG) program provide either user-friendly visualizations or ways to impute knowledge about potential targets using machine learning strategies. Watch these videos for step-by-step descriptions on how to use each web-based tool to accelerate illumination in under-studied proteins.
 
 ## Website Features
 
