### 2022-03-02 795a5a3
Update history.md
fix path to image
