### 2022-03-02 7e8855c
Update history.md
replace numbered list with bullets
