+++ docs/Release-Notes/history.md
@@ -4,58 +4,58 @@ Below is a list past CFDE workshops with links to workshop resources.
 
 ## 2022
 
-1. [Introduction to AWS with Human Microbiome Project (HMP) Data](https://hackmd.io/9EWhx5vYR3SDmd1Nl9rpyQ?view)  _February 22, 2022_ ([Watch the video recording](https://video.ucdavis.edu/media/Hackathon+-+Day+2+-+HMP/1_evaoqlpg))
-1. [CFDE Portal Demo](https://app.nih-cfde.org/) _February 21, 2022_ ([Watch the video recording](https://video.ucdavis.edu/media/Hackathon+-+Day+1+-+Portal+Demo/1_cr0ert97))
-1. [RNA-Seq Concepts, Design and Workflows](https://osf.io/kj5av/) _February 15, 2022_
-1. [Introduction to AWS for Cloud Computing](https://github.com/nih-cfde/training-and-engagement/wiki/A-Hands-on-Introduction-to-AWS-for-Cloud-Computing-January-26,-2022) _January 26, 2022_ 
-1. [Introduction to UNIX for Cloud Computing](https://github.com/nih-cfde/training-and-engagement/wiki/An-Introduction-to-UNIX-for-Remote-Computing:-January-19,-2022) _January 19, 2022_    
+* [Introduction to AWS with Human Microbiome Project (HMP) Data](https://hackmd.io/9EWhx5vYR3SDmd1Nl9rpyQ?view)  _February 22, 2022_ ([Watch the video recording](https://video.ucdavis.edu/media/Hackathon+-+Day+2+-+HMP/1_evaoqlpg))
+* [CFDE Portal Demo](https://app.nih-cfde.org/) _February 21, 2022_ ([Watch the video recording](https://video.ucdavis.edu/media/Hackathon+-+Day+1+-+Portal+Demo/1_cr0ert97))
+* [RNA-Seq Concepts, Design and Workflows](https://osf.io/kj5av/) _February 15, 2022_
+* [Introduction to AWS for Cloud Computing](https://github.com/nih-cfde/training-and-engagement/wiki/A-Hands-on-Introduction-to-AWS-for-Cloud-Computing-January-26,-2022) _January 26, 2022_ 
+* [Introduction to UNIX for Cloud Computing](https://github.com/nih-cfde/training-and-engagement/wiki/An-Introduction-to-UNIX-for-Remote-Computing:-January-19,-2022) _January 19, 2022_    
 
 ## 2021
 
-1. [An Introduction to AWS for Cloud Computing](https://github.com/nih-cfde/training-and-engagement/wiki/A-Hands-On-Introduction-to-AWS:-December-15,-2021)  _December 15, 2021_
-1. [An Introduction to UNIX for Cloud Computing](https://github.com/nih-cfde/training-and-engagement/wiki/An-Introduction-to-UNIX-for-Cloud-Computing---December-8,-2021)  _December 8, 2021_
-1. [Introduction to the UNIX Command Line](https://github.com/nih-cfde/training-and-engagement/wiki/An-Introduction-to-UNIX:-November-17,-2021) _November 17, 2021_  
-1. [Introduction to AWS for Cloud Computing](https://github.com/nih-cfde/training-and-engagement/wiki/A-Hands-On-Introduction-to-AWS:-November-10,-2021) _November 10, 2021_  
-1. [Introduction to R](https://github.com/nih-cfde/training-and-engagement/wiki/Introduction-to-R:-October-13,-2021) _October 13, 2021_  
-1. [Introduction to Conda](https://github.com/nih-cfde/training-and-engagement/wiki/A-Hands-On-Introduction-to-Conda:-October-6,-2021) _October 6, 2021_  
[…]
