### 2022-03-28 #132 @choldgraf
Hey all - just realized I didn't reply to this. I don't believe I am the right person to review a paper like this, and sadly am swamped for time right now anyway. Have you considered posting in the JupyterHub Team Compass? I believe that @manics might have some experience with AWS.
