### 2022-03-28 #132 @manics
I might be able to help... Can someone give me a TLDR of what to do/where to start before I commit? 😀
