### 2022-03-30 #132 @labarba
hi @manics — thank you! Have a look at the documentation here: https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html

If you agree to review, we'll start a new issue, which will have a review checklist at the top for each of the reviewers. You can tick off items there, and open issues on the target repository for things you would like to see improved, and we have a conversation in the review issue until all are happy. Sound good?
