### 2022-03-31 #167 @whedon
[REVIEW]: A virtual training module for introducing the use of Amazon Web Services
**Submitting author:** @abhijna (<a href="http://orcid.org/0000-0002-5861-3737">Abhijna Parigi</a>)
**Repository:** <a href="https://github.com/nih-cfde/training-and-engagement" target ="_blank">https://github.com/nih-cfde/training-and-engagement</a>
**Version:** 2022.07-aws-jose

- [x] **Version:** 2022.07-aws-jose

- [x] **Version:** 2022.07-aws-jose
