### 2022-03-31 #167 @labarba
@giuliaguidi, @manics — Thank you for agreeing to review for JOSE! This is where the action happens: work your way through the review checklist, feel free to ask questions or post comments here, and also open issues in the submission repository as needed. Godspeed!

Your Editor, @willingc, will take over the shepherding from here.
