### 2022-04-01 47fdc04
Update auto_assign_stable.yml
