+++ .github/actions/auto_assign_stable.yml
@@ -7,9 +7,6 @@ addAssignees: false
 # A list of reviewers to be added to pull requests (GitHub user name)
 reviewers:
   - raynamharris
-  - ACharbonneau
-  - ctb
-  - jeremywalter
 
 # A number of reviewers added to the pull request
 # Set 0 to add all the reviewers (default: 0)
@@ -22,7 +19,7 @@ numberOfReviewers: 0
 # A number of assignees to add to the pull request
 # Set to 0 to add all of the assignees.
 # Uses numberOfReviewers if unset.
-# numberOfAssignees: 2
+# numberOfAssignees: 1
 
 # A list of keywords to be skipped the process that add reviewers if pull requests include it
 # skipKeywords:
