### 2022-04-01 31bdd31
Update auto_assign_stable.yml
