+++ .github/actions/auto_assign_stable.yml
@@ -7,6 +7,7 @@ addAssignees: false
 # A list of reviewers to be added to pull requests (GitHub user name)
 reviewers:
   - raynamharris
+  - fp9695253
 
 # A number of reviewers added to the pull request
 # Set 0 to add all the reviewers (default: 0)
