+++ .github/actions/auto_assign_dev.yml
@@ -1,39 +1,24 @@
 # Set to true to add reviewers to pull requests
 addReviewers: true
 
-# Set to true to add assignees to pull requests
 addAssignees: author
 
 # Set to true to add reviewers from different groups to pull requests
-useReviewGroups: true
+useReviewGroups: false
 
-# A list of reviewers, split into different groups, to be added to pull requests (GitHub user name)
-reviewGroups:
-  groupA:
-    - raynamharris
-    - ACharbonneau
-    - ctb
-
-  groupB:
-    - jeremywalter
-
+# A list of reviewers to be added to pull requests (GitHub user name)
+reviewers:
+  - raynamharris
+  - fp9695253
 
 # A number of reviewers added to the pull request
 # Set 0 to add all the reviewers (default: 0)
 numberOfReviewers: 0
 
-# A list of assignees, overrides reviewers if set
-# assignees:
-#   - assigneeA
-
-# A number of assignees to add to the pull request
-# Set to 0 to add all of the assignees.
-# Uses numberOfReviewers if unset.
-# numberOfAssignees: 2
 
 # A list of keywords to be skipped the process that add reviewers if pull requests include it
-# skipKeywords:
-#   - wip
+skipKeywords:
+  - wip
 
 # The action will run only if the PR meets the specified filters
 #filterLabels:
