+++ docs/Release-Notes/index.md
@@ -1,0 +1,12 @@
+---
+layout: page
+title: April 2022 Release
+---
+
+# Latest Release
+
+**Updated April 15, 2022**
+
+## Website Features
+
+In this release, we added a new "Workshop History" page under the Release Notes tab that contains a list of all workshops taught by the CFDE training and coordination team. This page links to workshop resources and an image summarizing the total number of workshops and participants. We hope this makes it easier to find information from previous workshops. If you think a workshop is missing, please, please [email the training team](mailto:CFDE-HelpDesk%20<training@cfde.atlassian.net) or [open an issue](https://github.com/nih-cfde/training-and-engagement/issues/new).
