### 2022-04-06 44009b0
Update auto_assign_stable.yml
