+++ .github/actions/auto_assign_stable.yml
@@ -6,7 +6,10 @@ addAssignees: false
 
 # A list of reviewers to be added to pull requests (GitHub user name)
 reviewers:
-  - nih-cfde/training
+  - raynamharris
+  - jeremywalter
+  - jessicalumian
+  - ctb
 
 # A number of reviewers added to the pull request
 # Set 0 to add all the reviewers (default: 0)
