+++ README.md
@@ -24,4 +24,4 @@ Hello, if you'd like to contribute to the CFDE Training website, read on!
 
 By contributing to this repository, you agree to release all your contributions under the same terms as the license itself: [the Creative Commons Zero (aka Public Domain) license](./license.md). If you are OK with this condition, then we welcome both you and your contribution!
 
-If you have any questionsg, please [open an issue](https://github.com/nih-cfde/training-and-engagement/issues/new) or [email the training team](mailto:training@cfde.atlassian.net) and we will lend a hand.
+If you have any questions, please [open an issue](https://github.com/nih-cfde/training-and-engagement/issues/new) or [email the training team](mailto:training@cfde.atlassian.net) and we will lend a hand.
