### 2022-04-19 #167 @manics
I've read the paper, I'm going through the code/course
