### 2022-04-21 #167 @manics
See
- [ ] https://github.com/nih-cfde/training-and-engagement/issues/531
for my feedback on the source repository

@abhijna In the `Learning Objectives` section of the current PDF https://raw.githubusercontent.com/openjournals/jose-papers/jose.00167/jose.00167/10.21105.jose.00167.pdf

![image](https://user-images.githubusercontent.com/1644105/164428933-112dc351-b6cc-461a-99be-67f3d59cfa51.png)

- Is this formatting correct? Is it meant to be a bulletted list?
- The `online lesson` link to https://training.nih-cfde.org/en/latest/Cloud-Platforms/Introduction_to_Amazon_Web_Services/introtoaws1.html is broken, should it be https://training.nih-cfde.org/en/latest/Cloud-Platforms/Introduction_to_Amazon_Web_Services/introtoaws1/ ? More generally do you have a policy of maintaining deep-links long-term?
