+++ docs/Bioinformatic-Analyses/GWAS-in-the-cloud/aws_instance_setup.md
@@ -8,7 +8,7 @@ title: Set up an AWS Instance
 Setting up an AWS Instance
 ==========================
 
-Amazon offers a cloud computing platform called Amazon Web Services (AWS). Please check out our [full lessons on AWS](../../Cloud-Platforms/Introduction_to_Amazon_Web_Services/introtoaws1.md) for more details!
+Amazon offers a cloud computing platform called Amazon Web Services (AWS). Please check out our [full lessons on AWS](../../Cloud-Platforms/Introduction-to-AWS/index.md) for more details!
 
 AWS is not free, however, you can use the Free Tier for 12 months after you sign up for an AWS account. If you are no longer eligible for the Free Tier, you're charged at the [standard billing rate](https://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/free-tier-eligibility.html) for your AWS usage.
 
