### 2022-05-17 e5a25ad
set R landing page
