+++ docs/General-Tools/R-for-RNA-Seq/index.md
@@ -1,0 +1,56 @@
+---
+layout: page
+title: R for RNA-Seq Analysis
+hide:
+  - toc
+---
+
+An Introduction to R For RNA-Sequencing Analysis
+================================================
+
+RNA-Sequencing (RNA-Seq) is a popular method for
+determining the presence and quantity of RNA in biological samples. In
+this 3 hour workshop, we will use R to explore publicly-available
+RNA-Seq data from the [Gene Expression Tissue Project
+(GTEx)](https://gtexportal.org/home/). Attendees will be introduced to
+the R syntax, variables, functions, packages, and data structures common
+to RNA-Seq projects. We will use RStudio to import, tidy, transform, and
+visualize RNA-Seq count data. Attendees will learn tips and tricks for
+making the processes of data wrangling and data harmonization more
+manageable. This workshop will not cover cloud-based workflows for
+processing RNA-seq reads or statistics and modeling because these topics
+are covered in our [RNA-Seq Concepts](https://osf.io/kj5av/) and
+[RNA-Seq in the
+Cloud](https://github.com/nih-cfde/rnaseq-in-the-cloud/blob/stable/rnaseq-workflow.pdf)
+workshops. Rather, this workshop will focus on general R concepts
+applied to RNA-Seq data. 
+
+| Lesson name | Lenth
+| --- | 
+| [Introduction](./r4rnaseq-1.md) | 25 min
+| [Importing data](./r4rnaseq-2.md) | 30 min
+| [Visualizing data](./r4rnaseq-3.md) | 30 min
+| [Tidying and transforming d](./r4rnaseq-4.md) | 90 min
+| [Wrap-up](./r4rnaseq-5.md) | 5 min
+
+!!! note "Learning Objectives"
+
+In this workshop, you will learn how to use R and RStudio to:
+
+-   import and view files commonly associated with RNA-sequencing
+    experiments
+-   select variables and observations that are relevant to research
+    questions (tidy)
+-   create and rename variables (transform)
+-   join data frames by common variables (harmonize)
+-   visualize data using bar graphs, scatter plots, and box plots
+
+
+=== "Prerequisites"
+
+Familiarity with R and RNA Sequencing is not required but would be
+useful.
+
+=== "Tutorial Resources"
+
+Please refer to the [RStudio cheat sheets](https://www.rstudio.com/resources/cheatsheets/) for commonly used R functions!
