+++ docs/General-Tools/R-for-RNA-Seq/r4rnaseq-2.md
@@ -1,0 +1,332 @@
+## Importing and viewing data
+
+Data can be imported using packages from base R or the tidyverse. What
+are some differences between the data objects imported by base R
+functions such as `read.csv()` and Tidyverse functions such as
+`read_csv()`? To begin with, `read.csv()` replaces spaces and dashes
+periods in column names, and it also preserves row.names. On the other
+hand, `read_csv()` preserves spaces and dashes in column names but drops
+row.names. For this workshop, we will use `read_csv()`, which means we
+may have to replace dashes with periods so that our sample names in all
+objects with sample name information.
+
+Today, I will show you how to import the following files:
+
+1.  data/samples.csv
+2.  data/GTExHeart_20-29_vs_70-79.tsv
+3.  data/colData.HEART.csv
+4.  data/countData.HEART.csv.gz
+
+Later, you can practice on your own using the following files:
+
+1.  data/GTExMuscle_20-29_vs_70-79.tsv
+2.  data/colData.MUSCLE.csv
+3.  data/countData.MUSCLE.csv.gz
+
+The `GTExPortal.csv` file in `./data/` contains information about all
+the samples in the GTEx portal. Let’s import this file using
+`read.csv()`.
+
+``` r
+samples <- read.csv("./data/samples.csv")
+```
+
+After importing a file, there are multiple ways to view the data.
+`head()` to view the first few lines of each file. `names()` will print
+just the column names. `str` will compactly displaying the internal
+structure. `summary` will compute statistics.
+
+``` r
+#View(samples)
+head(samples)
+```
+
+    ##       SUBJID    SEX   AGE    DTHHRDY                   SAMPID           SMTS
+    ## 1 GTEX-1117F Female 60-69 Slow death GTEX-1117F-0226-SM-5GZZ7 Adipose Tissue
+    ## 2 GTEX-1117F Female 60-69 Slow death GTEX-1117F-0426-SM-5EGHI         Muscle
+    ## 3 GTEX-1117F Female 60-69 Slow death GTEX-1117F-0526-SM-5EGHJ   Blood Vessel
+    ## 4 GTEX-1117F Female 60-69 Slow death GTEX-1117F-0626-SM-5N9CS   Blood Vessel
+    ## 5 GTEX-1117F Female 60-69 Slow death GTEX-1117F-0726-SM-5GIEN          Heart
+    ## 6 GTEX-1117F Female 60-69 Slow death GTEX-1117F-1326-SM-5EGHH Adipose Tissue
+    ##   SMNABTCH  SMNABTCHD SMGEBTCHT SMAFRZE SMCENTER SMRIN SMATSSCR
+    ## 1 BP-43693 2013-09-17 TruSeq.v1  RNASEQ       B1   6.8        0
+    ## 2 BP-43495 2013-09-12 TruSeq.v1  RNASEQ       B1   7.1        0
+    ## 3 BP-43495 2013-09-12 TruSeq.v1  RNASEQ       B1   8.0        0
+    ## 4 BP-43956 2013-09-25 TruSeq.v1  RNASEQ       B1   6.9        1
+    ## 5 BP-44261 2013-10-03 TruSeq.v1  RNASEQ       B1   6.3        1
+    ## 6 BP-43495 2013-09-12 TruSeq.v1  RNASEQ       B1   5.9        1
+
+``` r
+tail(samples)
+```
+
+    ##          SUBJID    SEX   AGE         DTHHRDY                   SAMPID
[…]
