+++ docs/General-Tools/R-for-RNA-Seq/r4rnaseq-4.md
@@ -1,0 +1,602 @@
+## Tidying and Transforming Data
+
+[Data wrangling](https://en.wikipedia.org/wiki/Data_wrangling) is the
+process of tidying and transforming data to make it more appropriate and
+valuable for a variety of downstream purposes such as analytics. The
+goal of data wrangling is to assure quality and useful data. Data
+analysts typically spend the majority of their time in the process of
+data wrangling compared to the actual analysis of the data.
+
+**Tidying** your data means storing it in a consistent form. When your
+data is tidy, each column is a variable, and each row is an observation.
+Tidy data is important because the consistent structure lets you focus
+your struggle on questions about the data, not fighting to get the data
+into the right form for different functions. Some tidying functions
+include `pivot_longer()`, `pivot_wider()`, `separate()`, `unite()`,
+`drop_na()`, `replace_na()`. The “lubridate” package has a number of
+functions for tidying dates. You may also use `mutate()` function to
+convert objects from, say, characters or integers to factors or rename
+observations and variables.
+
+**Transforming** your data includes narrowing in on observations of
+interest (like all people in one city, or all data from the last year),
+creating new variables that are functions of existing variables (like
+computing speed from distance and time), and calculating a set of
+summary statistics (like counts or means). Summary functions such as
+`summarize()` and `count()` to create new tables with statistics. Before
+summarizing or counting a whole data frame, you can use `group_by()` to
+group variables. You can use `filter()` and `select()` to isolate
+specific rows or columns, respectively. If you want to sort columns,
+`arrange()` and `arrange(desc())` are two functions to familiarize
+yourself with.
+
+**Combining tables** can be accomplished in one of two ways. If all the
+columns or all the rows have all the same names, you can use `rbind()`
+or `cbind()`, respectively, to join the data frames. If however, each
+data frame have a column (or multiple columns) that contain unique
+identifiers, then you can use the family of join functions
+(`inner_join()`, `outter_join()`, `left_join()`, and `right_join()`)
+
+For each downstream analysis, you will likely use a series of tidying
+and transforming steps in various order to get your data in the
+appropriate format. Interest of creating dozens of intermediate files
+after each step, we will use the `%>%` operator to “pipe” the output of
+one function to the input of the other.
+
+Instead of going into each function or each process in detail in
+isolation, let’s start with some typical research questions and then
+piece together R functions to get the desired information
+
+### Filtering Data
+
+Filter is done in a few different ways depending on the type of
+variable. You can use `>` and less `<` to filter greater or less than a
+number. `==` and `!=` are used to filter by characters or factors that
+match or do not match a specific pattern. `%in% c()` is used to filter
[…]
