+++ docs/General-Tools/R-for-RNA-Seq/r4rnaseq-5.md
@@ -1,0 +1,36 @@
+## Wrap-Up
+
+### R Markdown
+
+The workshop notes for using this repository to teach an Introduction to
+R for RNA-seq are crated with the file `r4rnaseq-workshop.Rmd`.
+
+### References
+
+-   [R for Data Science by Hadley Wickham and Garrett
+    Grolemund](https://r4ds.had.co.nz/index.html)
+-   [Rouillard et al. 2016. The Harmonizome: a collection of processed
+    datasets gathered to serve and mine knowledge about genes and
+    proteins. Database
+    (Oxford).](http://database.oxfordjournals.org/content/2016/baw100.short)
+
+### Additional Resources
+
+-   [RStudio cheatsheet for
+    readr](https://raw.githubusercontent.com/rstudio/cheatsheets/master/data-import.pdf)
+-   [RStudio cheatsheet for
+    dplyr](https://raw.githubusercontent.com/rstudio/cheatsheets/master/data-transformation.pdf)
+-   [RStudio cheatsheet for data Wrangling with
+    dplyr](https://www.rstudio.com/wp-content/uploads/2015/02/data-wrangling-cheatsheet.pdf)
+-   [ggplot point
+    shapes](http://www.sthda.com/english/wiki/ggplot2-point-shapes)
+-   [Angus 2019 Intro to R
+    Lesson](https://angus.readthedocs.io/en/2019/R_Intro_Lesson.html)
+-   [Angus 2019 Differential Gene Expression in R
+    Lesson](https://angus.readthedocs.io/en/2019/diff-ex-and-viz.html)
+-   [Software Carpentry R
+    Lesson](http://swcarpentry.github.io/r-novice-inflammation/)
+
+*Note: the source document
+[r4rnaseq-workshop.Rmd](https://github.com/nih-cfde/training-rstudio-binder/blob/data/GTEx/r4rnaseq-workshop.Rmd)
+was last modified 3 May, 2022.*
