+++ docs/General-Tools/R-for-RNA-Seq/r4rnaseq-3.md
@@ -1,0 +1,220 @@
+## Visualizing data 
+
+`ggplot2` is a very popular package for making visualization. It is
+built on the “grammar of graphics”. Any plot can be expressed from the
+same set of components: a data set, a coordinate system, and a set of
+“geoms” or the visual representation of data points such as points,
+bars, line, or boxes. This is the template we build on:
+  
+  ```
+  ggplot(data = <DATA>, aes(<MAPPINGS>)) +
+  <geom_function>() +
+  ```
+
+We just used the `count()` function to calculate how many samples are in
+each group. The function for creating bar graphs (`geom_bar()`) also
+makes use of `stat = "count"` to plot the total number of observations
+per variable. Let’s use ggplot2 to create a visual representation of how
+many samples there are per tissue, sex, and hardiness.
+
+``` r
+# Visualizing data with ggplot2
+
+ggplot(samples, aes(x = SMTS)) +
+  geom_bar(stat = "count")
+```
+
+![](./images/bar1-1.png)
+
+In the last section, we will discuss how to modify the `themes()` to
+adjust the axes, legends, and more. For now, let’s flip the x and y
+coordinates so that we can read the sample names. We do this by adding a
+layer and the function `coord_flip()`
+
+``` r
+ggplot(samples, aes(x = SMTS)) +
+  geom_bar(stat = "count") + 
+  coord_flip()
+```
+
+![](./images/bar2-1.png)
+
+Now, there are two ways we can visualize another variable in addition to
+tissue. We can add color or we can add facets.
+
+Let’s first color the data by age bracket. Color is an aesthetic, so it
+must go inside the `aes()`. If you include `aes(color = AGE)` inside
+`ggplot()`, the color will be applied to every layer in your plot. If
+you add `aes(color = AGE)` inside `geom_bar()`, it will only be applied
+to that layer (which is important later when you layer multiple geoms.
+               
+               
+               
+``` r
+head(samples)
+ggplot(samples, aes(x = SMTS, color = AGE)) +
+geom_bar(stat = "count") + 
+coord_flip()
+```
+               
+![](./images/bar3-1.png)
+
+Note that the bars are outlined in a color according to hardy scale. If
+instead, you would the bars “filled” with color, use the aesthetic
+`aes(fill = AGE)`
+               
+``` r
+ggplot(samples, aes(x = SMTS, fill = AGE)) +
+geom_bar(stat = "count") + 
+coord_flip()
+```
+               
+![](./images/bar4-1.png)
+
+Now, let’s use `facet_wrap(~SEX)` to break the data into two groups
+based on the variable sex.
+               
+``` r
+ggplot(samples, aes(x = SMTS, fill = AGE)) +
+  geom_bar(stat = "count") + 
+  coord_flip() +
+  facet_wrap(~SEX)
+```
+               
+![](./images/bar5-1.png)
+
+With this graph, we have an excellent overview of the total numbers of
+RNA-Seq samples in the GTEx project, and we can see where we are missing
+data (for good biological reasons). However, this plot doesn’t show us
+Hardy Scale. It’s hard to layer 4 variables, so let’s remove Tissue as a
+variable by focusing just on one Tissue.
+               
+=== Challenge
+               
+    Create a plot showing the total number of samples per Sex, Age Bracket,
+    and Hardy Scale for *just* the Heart samples. Paste the code you used in
+    the chat.
+               
+=== Answer
+               
+     There are many options. Here are a few.
+               
+               ggplot(samples, aes(x = DTHHRDY, fill = AGE))  +
+                 geom_bar(stat = "count") +
+                 facet_wrap(~SEX) 
+               
[…]
