+++ docs/General-Tools/R-for-RNA-Seq/r4rnaseq-2.md
@@ -1,4 +1,4 @@
-## Importing and viewing data
+## Importing data
 
 Data can be imported using packages from base R or the tidyverse. What
 are some differences between the data objects imported by base R
