+++ docs/General-Tools/index.md
@@ -8,7 +8,7 @@ General Tools
 
 Tutorials in this section provide lessons for setting up compute environments, workflows, and websites. Helpful cheat sheets for commonly used commands are also available here.
 
-- [Snakemake Workflow Management](Snakemake/index.md)
 - [Introduction to Conda](Introduction-to-Conda/index.md)
-- [Building Websites](Web-Development/index.md)
-- [Cheat Sheets](Cheat-Sheets/index.md)
+- [R for RNA-Seq](R-for-RNA-Seq/index.md)
+- [Snakemake](Snakemake/index.md)
+- [Web Development](Web-Development/index.md)
