+++ docs/General-Tools/R-for-RNA-Seq/wrapup.md
@@ -1,4 +1,6 @@
-## Wrap-Up
+---
+title: Wrap-up
+---
 
 ### R Markdown
 
