+++ docs/General-Tools/R-for-RNA-Seq/wrapup.md
@@ -2,11 +2,8 @@
 title: Wrap-up
 ---
 
-### R Markdown
-
-The workshop notes for using this repository to teach an Introduction to
-R for RNA-seq are crated with the file   
-[`r4rnaseq-workshop.Rmd`](https://github.com/nih-cfde/training-rstudio-binder/blob/data/GTEx/r4rnaseq-workshop.Rmd).*
+Files for this workshop can be downloaded from GitHub at https://github.com/nih-cfde/training-rstudio-binder.
+Specifically, scripts and data are in the `GTEx` directory.
 
 ### References
 
@@ -16,14 +13,15 @@ R for RNA-seq are crated with the file
     datasets gathered to serve and mine knowledge about genes and
     proteins. Database
     (Oxford).](http://database.oxfordjournals.org/content/2016/baw100.short)
+-   [GTEx Data Portal](https://gtexportal.org/home/)
 
 ### Additional Resources
 
--   [RStudio cheatsheet for
+-   [RStudio cheat sheet for
     readr](https://raw.githubusercontent.com/rstudio/cheatsheets/master/data-import.pdf)
--   [RStudio cheatsheet for
+-   [RStudio cheat sheet for
     dplyr](https://raw.githubusercontent.com/rstudio/cheatsheets/master/data-transformation.pdf)
--   [RStudio cheatsheet for data Wrangling with
+-   [RStudio cheat sheet for data Wrangling with
     dplyr](https://www.rstudio.com/wp-content/uploads/2015/02/data-wrangling-cheatsheet.pdf)
 -   [ggplot point
     shapes](http://www.sthda.com/english/wiki/ggplot2-point-shapes)
