+++ docs/General-Tools/R-for-RNA-Seq/index.md
@@ -1,8 +1,6 @@
 ---
 layout: page
 title: Overview
-hide:
-  - toc
 ---
 
 An Introduction to R for RNA-Sequencing Analysis
@@ -35,7 +33,7 @@ applied to RNA-Seq data.
 
 ### Learning Objectives
 
-!!! In this workshop, you will learn how to use R and RStudio to: 
+!!! info In this workshop, you will learn how to use R and RStudio to: 
 
     - import and view files commonly associated with RNA-sequencing experiments
     - select variables and observations that are relevant to research questions (tidy)
