+++ docs/General-Tools/R-for-RNA-Seq/wrapup.md
@@ -32,3 +32,215 @@ Specifically, scripts and data are in the `GTEx` directory.
 -   [Software Carpentry R
     Lesson](http://swcarpentry.github.io/r-novice-inflammation/)
 
+### Appendix
+
+Below is a consolidated list of the commands used in this workshop. 
+
+```
+
+## Introduction
+
+2 + 2 * 100
+log10(0.05)
+
+pval <- 0.05
+pval
+
+-log10(pval)
+
+
+favorite_genes <- c("BRCA1", "JUN",  "GNRH1", "TH", "AR")
+favorite_genes
+
+#install.packages("ggplot2")
+
+library(ggplot2)
+library(tidyr)
+library(dplyr)
+
+
+## Importing Data
+
+samples <- read.csv("./data/samples.csv")
+
+#View(samples)
+head(samples)
+tail(samples)
+str(samples)
+summary(samples)
+
+
+counts <- read.csv("./data/countData.HEART.csv", row.names = 1)
+dim(counts)
+head(counts)[1:5]
+
+results <- read.table("./data/GTEx_Heart_20-29_vs_50-59.tsv")
+head(results)
+
+dim(samples)
+
+dplyr::count(samples, SMTS) 
+
+head(dplyr::count(samples, SMTS, SEX))
+
+head(dplyr::count(samples, SMTS, SEX, AGE, DTHHRDY ) )
+
+## Visualizing Data
+
+ggplot(samples, aes(x = SMTS)) +
+  geom_bar(stat = "count")
+
+
+ggplot(samples, aes(x = SMTS)) +
+  geom_bar(stat = "count") + 
+  coord_flip()
+
+
+ggplot(samples, aes(x = SMTS, color = AGE)) +
+  geom_bar(stat = "count") + 
+  coord_flip()
+
+
+ggplot(samples, aes(x = SMTS, fill = AGE)) +
+  geom_bar(stat = "count") + 
+  coord_flip()
+
+
+ggplot(samples, aes(x = SMTS, fill = AGE)) +
+  geom_bar(stat = "count") + 
+  coord_flip() +
+  facet_wrap(~SEX)
+
+
+ggplot(results, aes(x = logFC, y = -log10(adj.P.Val))) +
+  geom_point() 
+
+
+ggplot(results, aes(x = logFC, y = -log10(adj.P.Val))) +
+  geom_point() +
+  geom_hline(yintercept = -log10(0.05))
+
+
+ggplot(results, aes(x = logFC, y = -log10(adj.P.Val))) +
+  geom_point(aes(color = ifelse( adj.P.Val < 0.05, "p < 0.05", "NS"))) +
+  geom_hline(yintercept = -log10(0.05)) 
+
+
+ggplot(results, aes(x = logFC, y = -log10(adj.P.Val))) +
+  geom_point(aes(color = ifelse( adj.P.Val < 0.05, "p < 0.05", "NS"))) +
+  geom_hline(yintercept = -log10(0.05))  +
+  theme(legend.position = "bottom") +
+  labs(color = "20-29 vs 50-59 year olds", 
+       subtitle = "Heart Tissue Gene Expression")
+
+
+ggplot(samples, aes(x = SMCENTER, y = SMRIN)) +
+  geom_boxplot() +
+  geom_jitter(aes(color = SMRIN))
+
+
+## Wrangle Data
+
+results %>% 
+  filter(adj.P.Val < 0.05) %>% 
+  head()
+
+
+results %>% 
+  filter(logFC > 1 | logFC < -1) %>%
+  head()
+
+
+results %>% filter(adj.P.Val < 0.05,
+                   logFC > 1 | logFC < -1) %>%
+  arrange(adj.P.Val) %>%
+  head()
+
+
+
[…]
