### 2022-05-18 7ef9970
add ( ) to summary, extra space, and change and to an
