### 2022-05-18 b3dfd37
fix comaprisons, minor word changes
