+++ docs/General-Tools/R-for-RNA-Seq/import.md
@@ -6,7 +6,7 @@ Data can be imported using packages from base R or the tidyverse. What
 are some differences between the data objects imported by base R
 functions such as `read.csv()` and Tidyverse functions such as
 `read_csv()`? To begin with, `read.csv()` replaces spaces and dashes
-periods in column names, and it also preserves row.namesgit a. On the other
+periods in column names, and it also preserves the row names. On the other
 hand, `read_csv()` preserves spaces and dashes in column names but drops
 row.names. For this workshop, we will use `read_csv()`, which means we
 may have to replace dashes with periods so that our sample names in all
