+++ docs/General-Tools/R-for-RNA-Seq/import.md
@@ -2,15 +2,9 @@
 title: Import Data
 ---
 
-Data can be imported using packages from base R or the tidyverse. What
-are some differences between the data objects imported by base R
-functions such as `read.csv()` and Tidyverse functions such as
-`read_csv()`? To begin with, `read.csv()` replaces spaces and dashes
-periods in column names, and it also preserves the row names. On the other
-hand, `read_csv()` preserves spaces and dashes in column names but drops
-row.names. For this workshop, we will use `read_csv()`, which means we
-may have to replace dashes with periods so that our sample names in all
-objects with sample name information.
+Data can be imported using functions from `base` R (such as `read.csv()` and 
+`read.table()`) or with with functions from `readr`(such as `read_csv()` and `read_tsv()`). 
+There are subtle differences in the default behavior these function, included how they treat dashes and spaces in column names, whether headers and row names are default. For this workshop, we will use `read.csv()` and `read.table()`. 
 
 ### Files
 
@@ -28,7 +22,7 @@ Later, you can practice on your own using the following files:
 3.  data/countData.MUSCLE.csv.gz
 
 The `samples.csv` file in `./data/` contains information about all
-the samples in the GTEx portal. Let’s import this file using
+the samples in the GTEx portal v8. Let’s import this file using
 `read.csv()`.
 
 ### `read.csv()`
@@ -37,13 +31,15 @@ the samples in the GTEx portal. Let’s import this file using
 samples <- read.csv("./data/samples.csv")
 ```
 
-After importing a file, there are multiple ways to view the data.
-`head()` to view the first few lines of each file. `names()` will print
-just the column names. `str` will compactly displaying the internal
+### `head()` and `tail()`
+
+After importing a file, there are multiple ways to view the data. `head()` and `tail()` to view the first and last 6 lines of a file. 
+
+
+`str` will compactly displaying the internal
 structure. `summary()` will compute statistics.
 
 ``` r
-#View(samples)
 head(samples)
 ```
 
@@ -87,6 +83,11 @@ tail(samples)
     ## 1526        1
     ## 1527        1
     ## 1528        1
+
+### `str()` and `summary()`
+
+`str` will compactly displaying the internal structure. `summary` will compute statistics. 
+
 
 ``` r
 str(samples)
@@ -226,7 +227,7 @@ head(results)
     read.csv("./data/countData.MUSCLE.csv", row.names = 1) 
     ```
 
-### Summary Statistics
+### `dim()`
 
 You have now seen a variety of options for importing files. You may use
 many more in your R-based RNA-seq workflow, but these basics will get
@@ -241,6 +242,8 @@ dim(samples)
 ```
 
     ## [1] 1528   13
+    
+### `count()`    
 
 How many samples are there per tissue?
 
@@ -338,7 +341,6 @@ Now you have successfully imported data using multiple methods. Let's complete a
 | `read_csv()`          | A tidyR function for importing .csv files as tibbles            |
 | `read.table()`        | A base R function for importing tabular data with any delimiter |
 | `read_tsv()`          | A tidyR function for importing .tsv files as tibbles            |
-| `as_tibble()`         | Convert data frames to tibbles                                  |
 | `head()` and `tail()` | Print the first or last 6 lines of an object                    |
 | `dim()`               | A function that prints the dimensions of an object              |
 | `length()`            | Calculate the length of an object                               |
