### 2022-06-01 4c3555c
Update index.md
add notes and video
