+++ docs/General-Tools/UNIX/index.md
@@ -1,40 +1,28 @@
-# An Introduction to UNIX for Remote Computing 
+---
+layout: page
+title: Overview
+---
 
-**When:**  
+<div class="banner"><span class="banner-text">Lesson in Development</span></div>
 
-**Instructors:**  
 
-**Helpers:**   
+An Introduction to UNIX for Cloud Computing 
+================================================
 
-**Connect via**  
+This lesson introduces the UNIX command line to scientists and clinicians who need to use cloud-based and remote computers for basic and biomedical research. The commands used in today's workshop are commonly used by practitioners across disciplines. By practicing the commands, you will gain familiarity with the command line and confidence as a programmer. These commands can be used on a local or remote computer. We hope that you notice an improvement in the speed and reproducibility of your bioinformatics workflow. 
 
-**Computing environment:** [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/nih-cfde/training-rstudio-binder/data?urlpath=rstudio)
+The lesson materials were adapted from the UC Davis Data Lab's [Intro to Cloud Computing](https://ngs-docs.github.io/2021-august-remote-computing/) workshop, Data Carpentry's [Introduction to the Command Line for Genomics](https://datacarpentry.org/shell-genomics/) lesson, and the Lab for Data Intensive Biology's [Advanced Beginner/Intermediate Shell](https://dib-training.readthedocs.io/en/pub/2016-01-13-adv-beg-shell.html) workshop. 
+Please refer to our [UNIX Cheatsheet](https://training.nih-cfde.org/en/latest/General-Tools/Cheat-Sheets/bash_cheatsheet/) for a list of commonly used commands.
 
-:::warning
+### Learning objectives
 
-#### Questions: 
-
-1. What city/state/country are you in now?
-1. What time is it?
-1. Why did you hope to learn from this workshop? 
-:::
-
-#### Lesson Overview  
-
-This lesson introduces the UNIX command line to scientists and clinicians who need to use cloud-based and remote computers for basic and biomedical research. 
-
-These commands used in today's workshop are commonly used by practitioners across disciplines. By practicing the commands, you will gain familiarity with the command line and confidence as a programmer. These commands can be used on a local or remote computer. We hope that you notice an improvement in the speed and reproducibility of your bioinformatics workflow. 
-
-:::success
-
-#### Today's Learning Goals
+!!! info "In this workshop, you will learn to"
 
 * Understand basic UNIX command structure 
 * Navigate through hierarchical directory structures
 * Read, write, create, copy, move, and remove files
 * Understand wildcards and regular expression 
 * Redirect outputs and write for loops to build reproducible workflows
-:::
 
-The lesson materials were adapted from the UC Davis Data Lab's [Intro to Cloud Computing](https://ngs-docs.github.io/2021-august-remote-computing/) workshop, Data Carpentry's [Introduction to the Command Line for Genomics](https://datacarpentry.org/shell-genomics/) lesson, and the Lab for Data Intensive Biology's [Advanced Beginner/Intermediate Shell](https://dib-training.readthedocs.io/en/pub/2016-01-13-adv-beg-shell.html) workshop. 
-Please refer to our [UNIX Cheatsheet](https://training.nih-cfde.org/en/latest/General-Tools/Cheat-Sheets/bash_cheatsheet/) for a list of commonly used commands.
+**Click here to get started!** [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/nih-cfde/training-rstudio-binder/data?urlpath=rstudio)
+
