+++ docs/General-Tools/UNIX/natigate.md
@@ -1,0 +1,173 @@
+---
+title: Naviagate
+---
+
+### 
+
+UNIX commands are like sentences that can be very simple or complex. The simplest commands consist of only the command name. Many require the name of a file or directory and allow specially formatted arguments, known as flags or options, which modify the default behavior of the program. The grammar of a shell allows you to combine existing tools into powerful pipelines and handle large volumes of data automatically. Sequences of commands can be written into a script, improving the reproducibility of workflows. The ease of getting things done via the shell will increase with your exposure to the program.
+
+We should note that _folders_ are called **directories** at the command line. For all intents and purposes, they can be used interchangeably, but if you'd like more information please read about ["the folder metaphor"](https://en.wikipedia.org/wiki/Directory_%28computing%29#Folder_metaphor).
+
+This Binder comes preloaded with data provided by your instructors.  _If you want to do these exercises locally, on your own computer, you can [download the data here](https://s3.us-west-1.amazonaws.com/dib-training.ucdavis.edu/shell-data2.zip)._
+
+The commands `pwd` and `ls` are two simple commands that can be used to answer the two commonly asked questions "where am I?" and "what files are here?". We will use these frequently through the next sections.
+
+### `pwd`
+
+To answer the question "where am I?", we can use the **print working directory** or `pwd` command to see what directory we are currently located in. 
+
+```bash
+pwd
+```
+
+This will print the **absolute path** to the directory where we are located. An absolute path shows the complete series of directories you need to locate either a directory or a file starting from the **root directory** of your computer. The absolute path to the root directory is `/`. A useful way to start thinking about directories and files is through levels. At the highest level of your computer, you have the root directory. Everything that is contained in your computer is located in directories below your root directory. 
+
+The **home directory** is typically two levels down. For many Mac users, the home directory is `/Users/USERNAME`. If you are using the Binder provided for this workshop, the home directory is `/home/jovyan`. Because the absolute path to the home directory is different for every user, you can refer to the home directory with the tilde symbol `~`.
+
+```bash
+/home/jovyan
+```
+
+### `ls`
+
+The **list** or `ls` command is a simple yet powerful command that is used to list the contents of your computer. It can be executed with or without optional flags and directories or files. Let's look at the contents in our working directory by using the `ls`.
+
+```bash
+ls
+```
+
+We can see the following files:
+
+```bash
+books          images  README.md             seattle
+CFDE-logo.png  MiSeq   rstudio-terminal.png  
+```
+
+If we want more information about the files, such as the date they were created and their file size, we can add "flags" `-l` for long listing format.
+
+```bash
+ls -l
+```
+
+```bash
+drwxr-xr-x 2 jovyan jovyan   4096 Jan 18 21:13 books
+-rw-r--r-- 1 jovyan jovyan  71154 Jan 18 21:13 CFDE-logo.png
+drwxr-xr-x 2 jovyan jovyan   4096 Jan 18 21:13 images
[…]
