### 2022-06-01 cb960c8
Update navigate.md
formatting
