### 2022-06-01 ccd32d3
Create work.md
formatting
