+++ docs/General-Tools/UNIX/find.md
@@ -1,0 +1,178 @@
+---
+title: Find
+---
+
+
+A big part of data science is making sure what you _expect_ in a particular file is what you _have_ in that file. This is fairly easy when your files are small but is challenging when the files are much larger than your screen.
+
+
+To explore this topic in more detail, navigate to the `data/MiSeq/` directory.
+
+
+```bash
+cd ~/MiSeq
+ls
+```
+
+### FASTQ format
+
+This directory contains multiple [**FASTQ** files]((https://en.wikipedia.org/wiki/FASTQ_format)). A FASTQ file normally uses four lines per sequence.
+
+* Line 1 begins with a '@' character and is followed by a sequence identifier and an optional description (like a FASTA title line).
+* Line 2 is the raw sequence letters.
+* Line 3 begins with a '+' character and is optionally followed by the same sequence identifier (and any description) again.
+* Line 4 encodes the quality values for the sequence in Line 2, and must contain the same number of symbols as letters in the sequence.
+
+A FASTQ file containing a single sequence might look like this:
+
+
+::: info
+An example FASTQ file
+
+```
+@SEQ_ID
+GATTTGGGGTTCAAAGCAGTATCGATCAAATAGTAAATCCATTTGTTCAACTCACAGTTT
++
+!''*((((***+))%%%++)(%%%%).1***-+*''))**55CCF>>>>>>CCCCCCC65
+```
+:::
+
+
+
+We can use the `cat` command to print fastq files to the screen, but thousands of lines of text would crowd your screen. Instead, let's use the `head` command to view the first 8 lines file. You can copy the file name and paste it into the console or you can type and use tab complete to pick a particular file. 
+
+
+```bash
+head -n 4 F3D0_S188_L001_R1_001.fastq
+```
+
+```
+@M00967:43:000000000-A3JHG:1:1101:18327:1699 1:N:0:188
+NACGGAGGATGCGAGCGTTATCCGGATTTATTGGGTTTAAAGGGTGCGTAGGCGGCCTGCCAAGTCAGCGGTAAAATTGCGGGGCTCAACCCCGTACAGCCGTTGAAACTGCCGGGCTCGAGTGGGCGAGAAGTATGCGGAATGCGTGGTGTAGCGGTGAAATGCATAGATATCACGCAGAACCCCGATTGCGAAGGCAGCATACCGGCGCCCTACTGACGCTGAGGCACGAAAGTGCGGGGATCAAACAG
++
+AABABBFFFGGGGGGGGGGGGGGGGHHHHHHHGGGHHHHHGHGGGGGGGHGGGGGGHHHHHHHHHHGGGGGHHHHGHGGGGGGHHBGHGDGGGGGHHHGGGGHHHHHHHHGGGGGHG@DHHGHEGGGGGGBFGGEGGGGGGGG.DFEFFFFFFFDCFFFFFFFFFFFFFFFFFFFFFFFFFFDFDFFFEFFCFF?FDFFFFFFFFAFFFFFFFFFFFBDDFFFFFEFADFFFFFBAFFFA?EFFFBFF
+@M00967:43:000000000-A3JHG:1:1101:14069:1827 1:N:0:188
[…]
