### 2022-06-01 c5ef73f
Create redirect.md
hmm... doesn't actually cover > and >>
