### 2022-06-01 b4cfb39
Update redirect.md
add append
