### 2022-06-01 ba2a5e3
Update index.md
rm launch binder button
