### 2022-06-02 768411c
update ls output examples
