### 2022-06-02 3b42040
update head output, gunzip text file name
