### 2022-06-02 da7788d
fix head, remove success, fix UNIX commands
