### 2022-06-03 0c00165
Update read.md
key point formatting
