### 2022-06-03 206980b
update ls output, change format for Input/Output display
