+++ docs/General-Tools/UNIX/navigate.md
@@ -92,7 +92,7 @@ Let's return to our home directory using the `cd` command and a relative path, t
  
 Let's practice using the cd and ls commands to explore files in different directories.  
 
-Because books/ is in our working directory, we can navigate there with a relative path. What files are in the `books` directory and how large are they?
+Because `books/` is in our working directory, we can navigate there with a relative path. What files are in the `books` directory and how large are they?
 
 ```bash
 cd books/
@@ -133,31 +133,57 @@ We can see the following files:
      5. Yes, MiSeq is in the directory one level above.
      6. No, MiSeq is not in the directory two levels above.
      7. No, MiSeq is not in the root directory.
-:::
-
 
 Most, but not all of the files in the MiSeq directory are .fastq files. Which .fastq files are the largest? We can use the wildcard `*` to list only files that end in .fastq. We can use the `-S` option to sort by size.
 
+=== "Input"
 
-```bash
-cd ~/MiSeq
-pwd
-ls -lhS *.fastq
-```
+     cd ~/MiSeq
+     pwd
+     ls -lhS *.fastq
 
-```
--rwxr-xr-x 1 jovyan jovyan  11M Jan 18 21:13 F3D2_S190_L001_R1_001.fastq
--rwxr-xr-x 1 jovyan jovyan  11M Jan 18 21:13 F3D2_S190_L001_R2_001.fastq
--rwxr-xr-x 1 jovyan jovyan 9.2M Jan 18 21:13 F3D147_S213_L001_R1_001.fastq
--rwxr-xr-x 1 jovyan jovyan 9.2M Jan 18 21:13 F3D147_S213_L001_R2_001.fastq
--rwxr-xr-x 1 jovyan jovyan 7.1M Jan 18 21:13 F3D149_S215_L001_R1_001.fastq
--rwxr-xr-x 1 jovyan jovyan 7.0M Jan 18 21:13 F3D149_S215_L001_R2_001.fastq
--rwxr-xr-x 1 jovyan jovyan 6.7M Jan 18 21:13 F3D148_S214_L001_R1_001.fastq
--rwxr-xr-x 1 jovyan jovyan 6.7M Jan 18 21:13 F3D148_S214_L001_R2_001.fastq
--rwxr-xr-x 1 jovyan jovyan 4.3M Jan 18 21:13 F3D6_S194_L001_R1_001.fastq
--rwxr-xr-x 1 jovyan jovyan 4.3M Jan 18 21:13 F3D6_S194_L001_R2_001.fastq
-```
+=== "Expected Output"
 
+     -rwxr-xr-x 1 jovyan jovyan  11M May 26 03:05 F3D2_S190_L001_R1_001.fastq
+     -rwxr-xr-x 1 jovyan jovyan  11M May 26 03:05 F3D2_S190_L001_R2_001.fastq
+     -rwxr-xr-x 1 jovyan jovyan 9.2M May 26 03:05 F3D147_S213_L001_R1_001.fastq
+     -rwxr-xr-x 1 jovyan jovyan 9.2M May 26 03:05 F3D147_S213_L001_R2_001.fastq
+     -rwxr-xr-x 1 jovyan jovyan 7.1M May 26 03:05 F3D149_S215_L001_R1_001.fastq
[…]
