### 2022-06-03 0d26709
remove script mention (file not here and no context)
