+++ docs/General-Tools/UNIX/work.md
@@ -4,7 +4,7 @@ title: Work
 
 
 
-We are quite used to working with files using a GUI. In this section, you will learn how to copy, move, create, and delete directories and files.
+We are quite used to working with files using a graphical user interface (or GUI). In this section, you will learn how to copy, move, create, and delete directories and files.
 
 The `cp` and `mv` commands can be used to copy and move (or rename) files and directories respectively. For both commands, you must specify the old and new names. Specifying the path is necessary if you want to move files out of the current working directory.
 
@@ -16,7 +16,7 @@ mv [original-filename] [new-filename]
 
 ### `cp` 
 
-Let's make a copy of some raw data before we start modifying it. 
+Let's make a copy of some raw data before we start modifying it. We will use `ls` to check our work.
 
 ```bash
 cp book.txt book-copy.txt
@@ -25,7 +25,6 @@ ls
 
 ###  `mv`
 
-
 The `mv` command can be used to either move files to a new location or to rename them (which is essentially moving the contents from the old filename to the new file name. Let's use the `mv` command to rename the copied and compressed file back to the original name.
 
 ```bash
@@ -33,35 +32,26 @@ mv book-copy.txt book-2cities.txt
 
 ```
 
-Take care when renaming files. It is good practice to keep track of changes in file names and links to the source data. The commands used to get these books are stored in `get_books.sh`.  We will talk about how to execute this script later.
+!!! warning
 
-```bash
-head get_books.sh
-```
-
-
-:::warning
+Take care when renaming files. It is good practice to keep track of changes in file names and links to the source data. 
 
 ### `mkdir` 
 
-Now you know how to copy and move files, but you may encounter errors if you try to move files to a directory that doesn't exist. But, have no fear, we can create new directories at the command line with the command `mkdir` followed by the path to the directories you want to create. 
+Now you know how to copy and move files, but you may encounter errors if you try to move files to a directory that does not exist. But, have no fear, we can create new directories at the command line with the command `mkdir` followed by the path to the directories you want to create. 
 
 What happens when you run the following commands?
-
 
 ```bash
 mkdir data results images/
 mkdir -p data/results/images
 ```
-
-
 The first line creates two directories, results and images. The second line also create two directories, but they are nested with the parent directory data. 
-The `-p` argument creates parent directories if they don't already exist.
- 
+The `-p` argument creates parent directories if they do not already exist.
  
 ### `rmdir` 
 
-If you created some files or directories that you don't want, you can remove them with the `rm` and `rmdir` commands. 
+If you created some files or directories that you do not want, you can remove them with the `rm` and `rmdir` commands. 
 `rmdir` will only remove empty directories, but `rm -r` will remove recursively.
 
 ```bash
@@ -75,8 +65,6 @@ or
 rm -r data/results
 ```
 
-
-
 ### Key points
 
 | Command | Description |
@@ -86,4 +74,3 @@ rm -r data/results
 |rm [path] | removes (deletes) a file |
 |mkdir [path] | creates a new directory |
 |rmdir [path] | removees an empty  directory |
-:::
