### 2022-06-05 #167 @abhijna
Hi @manics, thanks for reviewing this paper!

> See
>
> * [ ]  [JOSE review feedback nih-cfde/training-and-engagement#531](https://github.com/nih-cfde/training-and-engagement/issues/531)
>   for my feedback on the source repository

> Have you considered adding a restrictive IAM policy? For instance you could allow just console access, ec2:*, and whatever else is required, along with a region restriction (it's annoying easy to accidentally switch region without realising).

Your concern about misuse risks to the AWS account holder is completely valid. I will add a note to the IAM Wiki about minimizing IAM user privileges for participants.

> Do you recommend people use https://training.nih-cfde.org/en/latest/Cloud-Platforms/Introduction_to_Amazon_Web_Services/introtoaws1/ directly, or would you prefer them to build their own site using your materials?

We recommend using our website and lesson materials. However, those who wish to co-opt this module are welcome to build their own lessons or re-use lessons from other sources.

> @abhijna In the `Learning Objectives` section of the current PDF https://raw.githubusercontent.com/openjournals/jose-papers/jose.00167/jose.00167/10.21105.jose.00167.pdf
>
> ![image](https://user-images.githubusercontent.com/1644105/164428933-112dc351-b6cc-461a-99be-67f3d59cfa51.png)
>
> * Is this formatting correct? Is it meant to be a bulletted list?

You are right, the learning objectives section is meant to be a bulleted list. I will fix that on the submission.

> * The `online lesson` link to https://training.nih-cfde.org/en/latest/Cloud-Platforms/Introduction_to_Amazon_Web_Services/introtoaws1.html is broken, should it be https://training.nih-cfde.org/en/latest/Cloud-Platforms/Introduction_to_Amazon_Web_Services/introtoaws1/ ? More generally do you have a policy of maintaining deep-links long-term?

Apologies for the broken link. It should be the latter. I will loop in Dr. Titus Brown  (senior author on this paper) to address the question of link maintenance.
