### 2022-06-05 3bae238
Fixed broken link
as per bullet #2 here: https://github.com/openjournals/jose-reviews/issues/167#issuecomment-1104981886
