### 2022-06-05 0398239
Learning objectives made bulleted list
as per review: https://github.com/openjournals/jose-reviews/issues/167#issuecomment-1104981886
