### 2022-06-05 cdd64ea
added clarity to re-use of online lessons
