+++ docs/Cloud-Platforms/Introduction-to-AWS/index.md
@@ -13,10 +13,10 @@ Amazon Web Services (AWS) is a subsidiary of Amazon that provides on-demand clou
 
 Est. time | Lesson name | Description
 --- | --- | ---
-30 mins | [Setting up an AWS instance](./introtoaws3.md) | How to start and configure an AWS instance?
-15 mins | [Connecting to an instance](./introtoaws4.md) | How to begin working on your AWS instance?
-20 mins | [Using the Screen Command](./introtoaws5_Screen.md) | How to run multiple screens and switch between tabs?
-10 mins | [Terminating an instance](./introtoaws5.md) | How to stop or terminate your AWS instance?
+30 mins | [Setting up an AWS instance](./aws1.md) | How to start and configure an AWS instance?
+15 mins | [Connecting to an instance](./aws2.md) | How to begin working on your AWS instance?
+20 mins | [Using the Screen Command](./aws3.md) | How to run multiple screens and switch between tabs?
+10 mins | [Terminating an instance](./aws4.md) | How to stop or terminate your AWS instance?
 
 !!! note "Learning Objectives"
 
@@ -36,6 +36,15 @@ Est. time | Lesson name | Description
 
 === "Tutorial Resources"
 
-    - [Vidlet: Setting up an AWS instance](./introtoaws2.md)
-
     - [Screen cheat sheet](../../General-Tools/Cheat-Sheets/screen_cheatsheet.md){:target="_blank"}
+
+
+Proceed directly to the [step-by-step tutorial](./aws1.md) or watch these 10-minute videos that demonstrate setting up and connecting to AWS instances via **Windows** or **Mac**.
+
+=== "Windows :fontawesome-brands-windows:"
+
+	<iframe id="kaltura_player" src="https://cdnapisec.kaltura.com/p/1770401/sp/177040100/embedIframeJs/uiconf_id/29032722/partner_id/1770401?iframeembed=true&playerId=kaltura_player&entry_id=0_1a78cmhv&flashvars[mediaProtocol]=rtmp&amp;flashvars[streamerType]=rtmp&amp;flashvars[streamerUrl]=rtmp://www.kaltura.com:1935&amp;flashvars[rtmpFlavors]=1&amp;flashvars[localizationCode]=en&amp;flashvars[leadWithHTML5]=true&amp;flashvars[sideBarContainer.plugin]=true&amp;flashvars[sideBarContainer.position]=left&amp;flashvars[sideBarContainer.clickToClose]=true&amp;flashvars[chapters.plugin]=true&amp;flashvars[chapters.layout]=vertical&amp;flashvars[chapters.thumbnailRotator]=false&amp;flashvars[streamSelector.plugin]=true&amp;flashvars[EmbedPlayer.SpinnerTarget]=videoHolder&amp;flashvars[dualScreen.plugin]=true&amp;flashvars[Kaltura.addCrossoriginToIframe]=true&amp;&wid=0_0rywna7h" width="550" height="500" allowfullscreen webkitallowfullscreen mozAllowFullScreen allow="autoplay *; fullscreen *; encrypted-media *" sandbox="allow-forms allow-same-origin allow-scripts allow-top-navigation allow-pointer-lock allow-popups allow-modals allow-orientation-lock allow-popups-to-escape-sandbox allow-presentation allow-top-navigation-by-user-activation" frameborder="0" title="Kaltura Player"></iframe>
+
+=== "macOS :fontawesome-brands-apple:"
+
[…]
