### 2022-06-06 ccfbd06
Delete aws2.md
moved to index
