+++ docs/Cloud-Platforms/Introduction-to-AWS/aws2.md
@@ -1,12 +1,0 @@
-# Setting up an AWS Instance Video Walk-through
-
-
-=== "Windows :fontawesome-brands-windows:"
-	You may choose to follow the walkthrough tutorial video first or proceed to follow the [step-by-step tutorial](./introtoaws3.md). This tutorial includes directions for connecting via **Windows**.
-
-	<iframe id="kaltura_player" src="https://cdnapisec.kaltura.com/p/1770401/sp/177040100/embedIframeJs/uiconf_id/29032722/partner_id/1770401?iframeembed=true&playerId=kaltura_player&entry_id=0_1a78cmhv&flashvars[mediaProtocol]=rtmp&amp;flashvars[streamerType]=rtmp&amp;flashvars[streamerUrl]=rtmp://www.kaltura.com:1935&amp;flashvars[rtmpFlavors]=1&amp;flashvars[localizationCode]=en&amp;flashvars[leadWithHTML5]=true&amp;flashvars[sideBarContainer.plugin]=true&amp;flashvars[sideBarContainer.position]=left&amp;flashvars[sideBarContainer.clickToClose]=true&amp;flashvars[chapters.plugin]=true&amp;flashvars[chapters.layout]=vertical&amp;flashvars[chapters.thumbnailRotator]=false&amp;flashvars[streamSelector.plugin]=true&amp;flashvars[EmbedPlayer.SpinnerTarget]=videoHolder&amp;flashvars[dualScreen.plugin]=true&amp;flashvars[Kaltura.addCrossoriginToIframe]=true&amp;&wid=0_0rywna7h" width="1000" height="500" allowfullscreen webkitallowfullscreen mozAllowFullScreen allow="autoplay *; fullscreen *; encrypted-media *" sandbox="allow-forms allow-same-origin allow-scripts allow-top-navigation allow-pointer-lock allow-popups allow-modals allow-orientation-lock allow-popups-to-escape-sandbox allow-presentation allow-top-navigation-by-user-activation" frameborder="0" title="Kaltura Player"></iframe>
-
-=== "macOS :fontawesome-brands-apple:"
-	You may choose to follow the walkthrough tutorial video first or proceed to follow the [step-by-step tutorial](./introtoaws3.md). This tutorial includes directions for connecting via **macOS**.
-
[…]
