### 2022-06-06 52ae1f4
Update and rename aws3.md to aws1.md
also update link to video
