### 2022-06-06 2eb9b5e
Update and rename aws6.md to aws4.md
add yml
