+++ docs/Cloud-Platforms/Introduction-to-AWS/aws4.md
@@ -1,7 +1,9 @@
-# Terminating an Instance
+---
+layout: page
+title: Terminating an Instance
+---
 
 Once you have completed your tasks and are sure you do not need the instance any longer, you may terminate the instance by returning to [AWS Management Console](https://us-west-1.console.aws.amazon.com/console/home?region=us-west-1){:target="_blank"}.
-
 
 !!! warning
 
