### 2022-06-06 449187d
Update aws1.md
update video wording
