### 2022-06-06 73aeba4
Update aws2.md
unhide toc
