### 2022-06-06 0731ea9
Update history.md
replace "video recording" with icon using <i class="fa fa-camera-retro fa-lg"></i>
