+++ docs/Release-Notes/history.md
@@ -11,16 +11,16 @@ Below is a list past CFDE workshops with links to workshop resources.
 
 ## 2022
 
-* [Introduction to R for RNA Sequencing Analysis](https://hackmd.io/MsWY1O9GQXGVDl2OmV2jxg)  _May 11, 2022_  [<i class="fa fa-video fa-lg"></i>](https://video.ucdavis.edu/media/CFDE+May+Hackathon+-+Intro+to+R+for+RNA-Seq/1_61sjku7o)
-* [CFDE Portal Demo with HMP Data, AWS, and sourmash](https://hackmd.io/dW9EoOh3T42eed9q21NLJQ?view#) _May 10, 2022_ [<i class="fa fa-video fa-lg"></i>](https://video.ucdavis.edu/media/t/1_yze086oo)
+* [Introduction to R for RNA Sequencing Analysis](https://hackmd.io/MsWY1O9GQXGVDl2OmV2jxg)  _May 11, 2022_  [<i class="fa fa-camera-retro fa-lg"></i>](https://video.ucdavis.edu/media/CFDE+May+Hackathon+-+Intro+to+R+for+RNA-Seq/1_61sjku7o)
+* [CFDE Portal Demo with HMP Data, AWS, and sourmash](https://hackmd.io/dW9EoOh3T42eed9q21NLJQ?view#) _May 10, 2022_ [<i class="fa fa-camera-retro fa-lg"></i>](https://video.ucdavis.edu/media/t/1_yze086oo)
 * [Introduction to R for RNA Sequencing Analysis](https://hackmd.io/SnorsWTbTTyRenptpjrhww?view) _April 27, 2022_ 
 * [Introduction to R for RNA Sequencing Analysis](https://hackmd.io/2ArmQGwGT0uUyL5Ehqy0hQ) _March 23, 2022_  
 * [RNA-Seq in the Cloud](https://osf.io/txcw8/) _March 7 and 9, 2022_  
 * [Introduction SigCom LINCS Search Engine](https://nih-cfde.github.io/2022-feb-hackathon/about/) _February 24, 2022_ 
 * [Extracellular RNA Communication (exRNA)](https://nih-cfde.github.io/2022-feb-hackathon/about/) _February 23, 2022_ 
 * [Introduction to to the Metabolomics Workbench](https://nih-cfde.github.io/2022-feb-hackathon/about/) _February 22, 2022_ 
-* [Introduction to AWS with Human Microbiome Project (HMP) Data](https://hackmd.io/9EWhx5vYR3SDmd1Nl9rpyQ?view)   _February 22, 2022_  [<i class="fa video fa-lg"></i>](https://video.ucdavis.edu/media/Hackathon+-+Day+2+-+HMP/1_evaoqlpg)
-* [CFDE Portal Demo](https://app.nih-cfde.org/) _February 21, 2022_  [ <i class="fa video fa-lg"></i>](https://video.ucdavis.edu/media/Hackathon+-+Day+1+-+Portal+Demo/1_cr0ert97)
+* [Introduction to AWS with Human Microbiome Project (HMP) Data](https://hackmd.io/9EWhx5vYR3SDmd1Nl9rpyQ?view)   _February 22, 2022_  [<i class="fa camera-retro fa-lg"></i>](https://video.ucdavis.edu/media/Hackathon+-+Day+2+-+HMP/1_evaoqlpg)
+* [CFDE Portal Demo](https://app.nih-cfde.org/) _February 21, 2022_  [ <i class="fa camera-retro fa-lg"></i>](https://video.ucdavis.edu/media/Hackathon+-+Day+1+-+Portal+Demo/1_cr0ert97)
[…]
