### 2022-06-06 42cd55f
Update history.md
fixed 2 links
