+++ docs/Release-Notes/history.md
@@ -19,8 +19,8 @@ Below is a list past CFDE workshops with links to workshop resources.
 * [Introduction SigCom LINCS Search Engine](https://nih-cfde.github.io/2022-feb-hackathon/about/) _February 24, 2022_ 
 * [Extracellular RNA Communication (exRNA)](https://nih-cfde.github.io/2022-feb-hackathon/about/) _February 23, 2022_ 
 * [Introduction to to the Metabolomics Workbench](https://nih-cfde.github.io/2022-feb-hackathon/about/) _February 22, 2022_ 
-* [Introduction to AWS with Human Microbiome Project (HMP) Data](https://hackmd.io/9EWhx5vYR3SDmd1Nl9rpyQ?view)   _February 22, 2022_  [<i class="fa camera-retro fa-lg"></i>](https://video.ucdavis.edu/media/Hackathon+-+Day+2+-+HMP/1_evaoqlpg)
-* [CFDE Portal Demo](https://app.nih-cfde.org/) _February 21, 2022_  [ <i class="fa camera-retro fa-lg"></i>](https://video.ucdavis.edu/media/Hackathon+-+Day+1+-+Portal+Demo/1_cr0ert97)
+* [Introduction to AWS with Human Microbiome Project (HMP) Data](https://hackmd.io/9EWhx5vYR3SDmd1Nl9rpyQ?view)   _February 22, 2022_  [<i class="fa fa-camera-retro fa-lg"></i>](https://video.ucdavis.edu/media/Hackathon+-+Day+2+-+HMP/1_evaoqlpg)
+* [CFDE Portal Demo](https://app.nih-cfde.org/) _February 21, 2022_  [<i class="fa fa-camera-retro fa-lg"></i>](https://video.ucdavis.edu/media/Hackathon+-+Day+1+-+Portal+Demo/1_cr0ert97)
 * [RNA-Seq Concepts, Design and Workflows](https://osf.io/kj5av/) _February 15, 2022_
 * [Introduction to AWS for Cloud Computing](https://github.com/nih-cfde/training-and-engagement/wiki/A-Hands-on-Introduction-to-AWS-for-Cloud-Computing-January-26,-2022) _January 26, 2022_ 
 * [Introduction to UNIX for Cloud Computing](https://github.com/nih-cfde/training-and-engagement/wiki/An-Introduction-to-UNIX-for-Remote-Computing:-January-19,-2022) _January 19, 2022_    
