### 2022-06-12 #167 @abhijna
Hi @manics,

I've fixed the broken links/render and incorporated all your suggestions. I have also added an optional Step 5b to the [IAM set up Wiki ](https://github.com/nih-cfde/training-and-engagement/wiki/IAM-account-for-AWS)on how to [add a permissions boundary](https://github.com/nih-cfde/training-and-engagement/wiki/IAM-account-for-AWS#optional-step-5b-adding-a-permissions-boundary).

I am unable to `@` tag Titus to this thread to comment on link maintenance.

@willingc, what is the next step in the review process?
