### 2022-06-13 #167 @arfon
@willingc – I think it's still @whedon in here... (@editorialbot is only active on JOSS thus far).
