### 2022-06-13 #167 @arfon
Ah OK. The authors need to make an archive of their work with Zenodo, figshare or somewhere else. See the docs about this here: https://openjournals.readthedocs.io/en/jose/editing.html#after-acceptance
