### 2022-06-14 29cd335
Create find-export.md
first part of the ihmp walkthrough lesson from the may hackathon  (see https://github.com/nih-cfde/2022-may-hackathon/issues/11 and https://github.com/nih-cfde/2022-may-hackathon/issues/10
