+++ docs/Common-Fund-Tools/CFDE-Portal/iHMP/find-export.md
@@ -1,0 +1,94 @@
+---
+layout: page
+title: Using the portal to find and export file information
+---
+
+
+#  Using the portal to find and export file information
+
+## Create a new personal collection
+
+Go to [the CFDE data portal at app.nih-cfde.org/](https://app.nih-cfde.org/). 
+
+Log in (upper right).
+
+![](https://i.imgur.com/YVXgMVK.png)
+
+Under your username (upper right), create a new personal collection. 
+
+
+![](https://i.imgur.com/D2eEXg2.png)
+
+
+For name, you can use "Tuesday demo" or anything else. You can leave description blank.
+
+![](https://i.imgur.com/eNoJFep.png)
+
+
+### Find some files
+
+Go back to [the CFDE data portal main page](https://app.nih-cfde.org/). 
+
+Select "File" (upper left).
+
+![](https://i.imgur.com/nIlZ2Jw.png)
+
+Use the facets on the left to select:
+* Common Fund Program: HMP
+* Project: "Longitudinal multi 'omics"
+* "has persistent ID" - True
+* Uncompressed size in Bytes - 50000000 to 60000000 (50 MB to 60 MB).
+
+![](https://i.imgur.com/7SAZK0X.png)
+
+
+![](https://i.imgur.com/9wOPGAY.png)
+
+
+With these selections, the first result should have "Filename" of `SRR5935743_1.fastq`.
+
+![](https://i.imgur.com/ULbqD7W.png)
+
+
+## Add files to your personal collection
+
+Click on the first result to get a detail view. Then add it to your personal collection:
+* scroll down to "Part of personal collection" and click "Link records."
+* Select your personal collection, click "Link" (upper right).
+
+![](https://i.imgur.com/76nE4vc.png)
+
+![](https://i.imgur.com/lWWBz0m.png)
+
+
+Click back in your browser, to get back to your filtered search.
+
+Repeat linking to a collection with the third result (Filename: `SRR5950647_1.fastq`). Add it to the same personal collection.
+
+![](https://i.imgur.com/aOX8Gz7.png)
+
+
+:::info
+For today, here are the direct links to the two files we'll be using:
+* [file 1, `SRR5935743_1.fastq`](https://app.nih-cfde.org/chaise/record/#1/CFDE:file/nid=528892)
+* [file 2, `SRR5950647_1.fastq`](https://app.nih-cfde.org/chaise/record/#1/CFDE:file/nid=531342)
+:::
+
+(Note, you could select any files you like, but these are small enough to work and I know what the results will be. So it's good for today's demo; I suggest trying new/different files as a Thursday exercise!)
+
+## Export your personal collection
+
+Go to your collection, and select "export" and choose "NCPI manifest format."
+
+:::info
+**What is NCPI?**
+"NCPI" stands for "NIH Cloud Platform Interoperability", an effort by the NIH to convene around interoperation for cloud workbenches.
+:::
+
+### Examine the NCPI manifest file
+
+You should now have a CSV file in your Downloads that, when examined with a spreadsheet program, looks like this:
+
+![](https://hackmd.io/_uploads/HkshxgdLq.png)
+
+The key piece of information in here is the `drs_uri` column, which provides a [Data Repository Service](https://ga4gh.github.io/data-repository-service-schemas/preview/release/drs-1.0.0/docs/) location from which to download the files.
