### 2022-06-14 nih-cfde/training-and-engagement#531 @manics
Thanks for responding to the review! I've got one more small suggestion, add a `CONTRIBUTING.md` file to the repo:
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors
