### 2022-06-14 10cecb2
Create CAVATICA.md
add importing files to cavatica with drs
