+++ docs/Common-Fund-Tools/CFDE-Portal/iHMP/CAVATICA.md
@@ -1,0 +1,36 @@
+---
+layout: page
+title: Adding data to CAVATICA with DRS IDs
+---
+
+# Adding data to CAVATICA with DRS IDs
+
+CAVATICA is a storage, sharing, and analysis platform designed by Seven Bridges.
+
+- [CAVATICA Portal](https://cavatica.sbgenomics.com/)
+- [CAVATICA Documentation](https://docs.cavatica.org/docs)
+
+To add data to CAVATICA, complete the following steps.
+
+Login to CAVATICA https://cavatica.sbgenomics.com/home
+
+  ![drs-1](../../images/drs-1.png)
+
+Click "Yes, I authorize" to allow CAVATICA to view your Researcher Auth Service (RAS) information
+
+![drs-2](../../images/drs-2.png)
+
+Create a new project by clicking "+ Create Project". Provide a name and billing group then click "Create".  
+
+Click the "Files" tab then click "+ Add files" drop down menu select "GA4GH Data Repository Service (DRS)".  
+
+![drs-3](../../images/drs-3.png)
+
+Enter the DRS identifiers (separated by a comma, enter, or tab) and click "Import"
+* _(Follow the steps in [the pentultimate lesson](./find-export) to get 2 DRS identifiers)_
+* drs://drs.hmpdacc.org/mZBm6TYnDQoS
+* drs://drs.hmpdacc.org/1CzGt5DyGwcOe
+
+Wait for the import to finish then click "Download" to download a raw file locally confirm that it is the correct format.
+
+![drs-4](../../images/drs-4.png)
