### 2022-06-14 d0b7674
Update index.md
add ihmp lessons
