### 2022-06-14 cfc1f4e
Update index.md
add column to highlight he use case groupings
