### 2022-06-14 6e31a4e
Create CONTRIBUTING.md
