### 2022-06-14 nih-cfde/training-and-engagement#531 @raynamharris
thank you for the suggestion @manics. I have started a PR to add a contributing file here https://github.com/nih-cfde/training-and-engagement/pull/542
