### 2022-06-14 9a0df4f
Update PULL_REQUEST_TEMPLATE.md
adds the reviewer checklist, some reorganization and rewording
