+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,19 +1,37 @@
 ## PR Checklist
 
 - [ ] Release label added
-- [ ] PR category label added
-- [ ] PR mergeable
+- [ ] Category label added
 - [ ] Clean build logs
 - [ ] Linked related issues
-- [ ] Colorblindness test https://www.toptal.com/designers/colorfilter/
+- [ ] Ready for review 
+- [ ] Ready to merge
 
 ## PR Description
-Briefly describe the changes this PR will add/address
-**Preview link**
-Add the RTD auto generated preview link
 
-## Review format
-Add how you would like to receive the review (direct editing, as comments, notes etc)
+Briefly describe the changes in this PR and the type of feedback desired.
 
-## Timeline
-Add expected timeline to finish review and merge adding note for time critical updates
+## Feedback Requested
+
+Include how you would like to receive the review (direct editing, as comments, notes, etc).
+Delete any irrelevant reviewer checks below.
+Add an expected timeline to finish the review and merge the PR.
+
+## Reviewer Checklist 
+
+- [ ] spelling and grammar
+- [ ] successful run of all installation and code chunks
+- [ ] sufficient explanation and details for the tutorial content
+- [ ] code syntax and naming convention
+- [ ] adherence of tutorial format to style guide
+- [ ] functional links (inter and intra)
+- [ ] accessible hyperlink text ([colorblindness test](https://www.toptal.com/designers/colorfilter))
+
+## PR Preview
+
+Add the Read the Docs auto-generated preview link.
+
+<!---
+Note: Will look something like this, where the number is the same as the PR number.
+https://cfde-training-and-engagement--535.com.readthedocs.build/en/535/
+--->
