### 2022-06-14 ae3328c
Delete comment_pr.yml
moved to PR template
