+++ .github/workflows/comment_pr.yml
@@ -1,28 +1,0 @@
-name: Reviewer checklist
-
-on:
-  pull_request:
-    branches: [dev]
-    types: [opened, reopened]
-
-jobs:
-  example_comment_pr:
-    runs-on: ubuntu-latest
-    name: Comment PR
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v1
-
-      - name: Comment PR
-        uses: thollander/actions-comment-pull-request@v1
-        with:
-          message: |
-            ## Reviewer Checklist 
-            - [ ] spelling and grammar
-            - [ ] successful run of all installation and code chunks
-            - [ ] sufficient explanation and details for the tutorial content
-            - [ ] code syntax and naming convention
-            - [ ] adherence of tutorial format to style guide
-            - [ ] functional links (inter and intra)
-            - [ ] accessible hyperlink text
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
