### 2022-06-15 9b4a6ec
reorder so that resources are on top of pre-reqs
