+++ docs/Website-Style-Guide/index.md
@@ -19,15 +19,14 @@ Time | Section | About
 10 mins | 6. [Pangeo binder style guide](./5PangeoBinderGuide.md)| How to create Pangeo binders?
 5 mins | 7. [Practice repo style guide](./6PracticeGithubRepos.md) | How to create a practice Github repo?
 
-=== "Prerequisites"
-
-    To contribute to the CFDE's website, you must be onboarded to the CFDE. Contact us at <training@cfde.atlassian.net> for help if you're interested in contributing training materials!
-
 === "Resources"
 
     - [Get started with new tutorials using the tutorial template markdown doc](https://github.com/nih-cfde/training-and-engagement/blob/dev/docs/CFDE-Internal-Training/Website-Style-Guide/tutorial_template_docs/TutorialTemplate.md)
     - [Powerpoint slide template](https://drive.google.com/drive/u/0/folders/14dOaf7-G4k7rCw5mL2Q5jdRWXrO0Y5i-)
 
+=== "Prerequisites"
+
+    To contribute to the CFDE's website, you must be onboarded to the CFDE. Contact us at <training@cfde.atlassian.net> for help if you're interested in contributing training materials!
 
 Location of important files in the training Github repo:
 ```
