+++ docs/Website-Style-Guide/index.md
@@ -3,10 +3,6 @@ layout: page
 title: Website Style Guide
 ---
 
-Website Style Guide
-
-===============================================================
-
 This is a style guide for content on the CFDE [training website](https://cfde-training-and-engagement.readthedocs-hosted.com/en/latest/).
 
 Time | Section | About
