+++ CONTRIBUTING.md
@@ -6,7 +6,7 @@ Frequently asked questions are posted on the [Discussions Board](https://github.
 
 The [Wiki](https://github.com/nih-cfde/training-and-engagement/wiki) contains lesson notes from workshops and resources for instructors and attendees. Only instructors and administrators should edit these pages. If you have questions or comments about the wiki, please email [support@cfde.atlassian.net](mailto:support@cfde.atlassian.net) or [open an issue](https://github.com/nih-cfde/training-and-engagement/issues/new).
 
-The `stable` branch of this repository builds the public [training website](https://training.nih-cfde.org) for the `[docs/](https://github.com/nih-cfde/training-and-engagement/tree/dev/docs)` directory. The `dev` branch contains updates that will be merged as part of the next [release](https://github.com/nih-cfde/training-and-engagement/releases). If you would like to submit a new lesson, please read the [website style guide](https://github.com/nih-cfde/training-and-engagement/blob/dev/docs/Website-Style-Guide/0index.md). To suggest changes or add files, fork the repo and open a pull request against the `dev` branch.
+The `stable` branch of this repository builds the public [training website](https://training.nih-cfde.org) for the `[docs/](https://github.com/nih-cfde/training-and-engagement/tree/dev/docs)` directory. The `dev` branch contains updates that will be merged as part of the next [release](https://github.com/nih-cfde/training-and-engagement/releases). If you would like to submit a new lesson, please read the [website style guide](https://training.nih-cfde.org/en/latest/Website-Style-Guide/). To suggest changes or add files, fork the repo and open a pull request against the `dev` branch.
 
 If you have training-related questions, contact us at [training@cfde.atlassian.net](mailto:training@cfde.atlassian.net). For all other questions, contact us at [support@cfde.atlassian.net](mailto:support@cfde.atlassian.net).
 
