### 2022-06-15 18a0869
Update index.md
move content from nested index files up a level
