+++ docs/Common-Fund-Tools/index.md
@@ -1,17 +1,99 @@
 ---
 layout: page
-title: Overview
+title: Common Fund Tools
 ---
-Common Fund Tools
-=======================
 
-Tutorials in this section provide lessons for finding data sets across the NIH Common Fund.
+## CFDE Search Portal
 
-- [CFDE Portal Tutorials](CFDE-Portal/index.md) (in development)
-- [Kids First Tutorials](Kids-First/index.md)
-- [Illuminating the Druggable Genome (IDG) Tutorials](IDG.md)
+The [NIH Common Fund (CF)](https://commonfund.nih.gov) has funded a wide variety of data types and studies that are of interest to clinical and biomedical researchers, however those datasets are hosted on an equally large number of websites, with varying query systems. The [Common Fund Data Ecosystem (CFDE) Portal](https://app.nih-cfde.org) is a unified system for searching across the entire CF portfolio in a single search, and is the first step in addressing the goal of making CF data more [Find-able, Accessible, Interoperable and Reusable (FAIR)](https://www.nih-cfde.org/product/fair-cookbook/). The wide range of data types, models and formats used by Common Fund Programs are being harmonized using well-defined metadata, common controlled vocabularies using the [Crosscut Metadata Model](https://www.nih-cfde.org/product/cfde-c2m2/).
+
+Est. Time | Use Case | Lesson name | Description
+--- | --- | --- | ---
+30 mins | 3 | [Using the CFDE Search Portal to Find Files](./iHMP/find-export.md) | Find iHMP data with DRS IDs
+10 mins | 3 | [Adding data to CAVATICA with DRS IDs](./iHMP/CAVATICA.md) | Add files to CAVATICA
+
+!!! note "Learning Objectives"
+
+    In this tutorial you will learn:
+
+    - How to search the CFDE Portal to find data
+    - How to use DRS IDs to access data
+
+=== "Prerequisites"
+
+      - Access to MacOS, Windows or Linux
+      - Access to the internet and a web browser
+      
+## Kids First  
+
+Kids First & Cavatica Lessons
+===================================
+
+The [NIH Common Fund-supported Gabriella Miller Kids First Data Resource
+Center](https://kidsfirstdrc.org/) (KFDRC) enables researchers, clinicians, and
+patients to work together to accelerate research and promote new
+discoveries for children affected with cancer and structural birth
+defects. Data from over 16,000 individuals, including DNA and RNA sequence data, is
+available to empower your research today. Data collected from more than
+30,000 samples are expected to be available in the next few years.
+
+Our aim is to teach researchers basic concepts, skills, and tools for
+working with data so that they can get more done in less time, and with
+less pain. The goal of this training module is to enable clinicians and
+researchers to use the data portal to build cohorts, and to move that data to an analysis platform ([Cavatica](https://cavatica.squarespace.com/)).
 
 
-You may also be interested in this lesson, which uses Common Fund program data:
+Setting up your KF Portal Permissions:
 
-- [RNA-Seq on Cavatica](../Bioinformatic-Analyses/RNAseq-on-Cavatica/rna_seq_1.md)
+Est. Time | Lesson name | Description
+--- | --- | ---
+5 mins | [The Setup](./Kids-First/Portal-Setup-And-Permissions/KF_2_The_Setup.md) | Introduction
+20 mins | [KF Portal Registration](./Kids-First/Portal-Setup-And-Permissions/KF_3_KF_Registration.md) | Register for KF portal access
[…]
