+++ docs/Common-Fund-Tools/index.md
@@ -3,14 +3,16 @@ layout: page
 title: Common Fund Tools
 ---
 
+# Common Fund Tools
+
 ## CFDE Search Portal
 
 The [NIH Common Fund (CF)](https://commonfund.nih.gov) has funded a wide variety of data types and studies that are of interest to clinical and biomedical researchers, however those datasets are hosted on an equally large number of websites, with varying query systems. The [Common Fund Data Ecosystem (CFDE) Portal](https://app.nih-cfde.org) is a unified system for searching across the entire CF portfolio in a single search, and is the first step in addressing the goal of making CF data more [Find-able, Accessible, Interoperable and Reusable (FAIR)](https://www.nih-cfde.org/product/fair-cookbook/). The wide range of data types, models and formats used by Common Fund Programs are being harmonized using well-defined metadata, common controlled vocabularies using the [Crosscut Metadata Model](https://www.nih-cfde.org/product/cfde-c2m2/).
 
-Est. Time | Use Case | Lesson name | Description
---- | --- | --- | ---
-30 mins | 3 | [Using the CFDE Search Portal to Find Files](./iHMP/find-export.md) | Find iHMP data with DRS IDs
-10 mins | 3 | [Adding data to CAVATICA with DRS IDs](./iHMP/CAVATICA.md) | Add files to CAVATICA
+Est. Time |  Lesson name | Description
+--- | --- | --- |
+30 mins | [Using the CFDE Search Portal to Find Files](./iHMP/find-export.md) | Find Human Microbiome Project (HMP) data with DRS IDs
+10 mins | [Adding files to CAVATICA with DRS IDs](./iHMP/CAVATICA.md) | Add files to CAVATICA using DRS IDs
 
 !!! note "Learning Objectives"
 
@@ -24,13 +26,11 @@ Est. Time | Use Case | Lesson name | Description
       - Access to MacOS, Windows or Linux
       - Access to the internet and a web browser
       
-## Kids First  
+## Kids First & Cavatica Lessons
 
-Kids First & Cavatica Lessons
-===================================
 
-The [NIH Common Fund-supported Gabriella Miller Kids First Data Resource
-Center](https://kidsfirstdrc.org/) (KFDRC) enables researchers, clinicians, and
+The [NIH Common Fund-supported Gabriella Miller Kids First (KF) Data Resource
+Center (DRC)](https://kidsfirstdrc.org/) enables researchers, clinicians, and
 patients to work together to accelerate research and promote new
 discoveries for children affected with cancer and structural birth
 defects. Data from over 16,000 individuals, including DNA and RNA sequence data, is
@@ -42,8 +42,7 @@ working with data so that they can get more done in less time, and with
 less pain. The goal of this training module is to enable clinicians and
 researchers to use the data portal to build cohorts, and to move that data to an analysis platform ([Cavatica](https://cavatica.squarespace.com/)).
 
-
-Setting up your KF Portal Permissions:
+### Setting up your KF Portal Permissions
 
 Est. Time | Lesson name | Description
 --- | --- | ---
@@ -53,7 +52,7 @@ Est. Time | Lesson name | Description
 20 mins | [Connecting Accounts](./Kids-First/Portal-Setup-And-Permissions/KF_5_ConnectingAccounts.md) | Connect KF and Cavatica
 20 mins | [Accessing Data](./Kids-First/Portal-Setup-And-Permissions/KF_6_AccessingData.md) | Set up data access authorization
 
-Exploring Data in the KF Portal:
+### Exploring Data in the KF Portal
 
[…]
