### 2022-06-15 5ba8dbd
fix table, combine learning objectives
