### 2022-06-15 f72d520
Update find-export.md
add intro, update formatting
