+++ docs/Common-Fund-Tools/CFDE-Portal/iHMP/find-export.md
@@ -3,8 +3,19 @@ layout: page
 title: Using the portal to find and export file information
 ---
 
+#  Using the CFDE Search portal to find and export file information
 
-#  Using the portal to find and export file information
+The Common Fund Data Ecosystem Coordinating Center (CFDE-CC) supports efforts to make Common Fund data sets more findable, accessible, interoperable, and reusable (FAIR) for the scientific community through collaboration, end-user training, and data set sustainability. The CFDE-CC manages and organizes CFDE activities, engages with participating Common Fund programs, connect with user communities, supports training, develops tools and standards, and provides technical expertise to Common Fund programs.
+
+The [CFDE Search Portal](https://app.nih-cfde.org/) uses the Crosscut Metadata Model (C2M2) a flexible metadata standard for describing experimental resources in biomedicine and related fields. This portal supports faceted search of metadata concepts such as anatomical location, species, and assay type, across a wide variety of datasets using a controlled vocabulary (we do not currently support protected metadata). This allows researchers to find a wide variety of data that would otherwise need to be searched individually, using varying nomenclatures. The portal only accepts C2M2 data packages from Common Fund Programs.
+
+This tutorial focuses on the Human Microbiome Project (HMP). The goal of this tutorial is to identify small FASTQ files with persistent identifiers from a longitudinal multi-comics study. Please refer to the [Portal User Guide](https://docs.nih-cfde.org/en/latest/about/portalguide/) for a detailed description of all portal features. 
+
+!!! note "Learning Objectives"
+    In this tutorial, you will learn how to
+    - access the CEFD search portal
+    - create personal collections
+    - search for files meeting a specific criterion
 
 ## Create a new personal collection
 
@@ -16,9 +27,7 @@ Log in (upper right).
 
 Under your username (upper right), create a new personal collection. 
 
-
 ![](https://i.imgur.com/D2eEXg2.png)
-
 
 For name, you can use "Tuesday demo" or anything else. You can leave description blank.
 
@@ -29,7 +38,7 @@ For name, you can use "Tuesday demo" or anything else. You can leave description
 
 Go back to [the CFDE data portal main page](https://app.nih-cfde.org/). 
 
-Select "File" (upper left).
+Select <span class="highlight_txt">File</span> (upper left).
 
 ![](https://i.imgur.com/nIlZ2Jw.png)
 
@@ -53,15 +62,15 @@ With these selections, the first result should have "Filename" of `SRR5935743_1.
 ## Add files to your personal collection
 
 Click on the first result to get a detail view. Then add it to your personal collection:
-* scroll down to "Part of personal collection" and click "Link records."
-* Select your personal collection, click "Link" (upper right).
+* scroll down to "Part of personal collection" and click  <span class="highlight_txt">Link records</span>.
+* Select your personal collection, click <span class="highlight_txt">Link</span> (upper right).
 
 ![](https://i.imgur.com/76nE4vc.png)
 
 ![](https://i.imgur.com/lWWBz0m.png)
 
 
-Click back in your browser, to get back to your filtered search.
+Click <span class="highlight_txt">back</span> in your browser, to get back to your filtered search.
 
 Repeat linking to a collection with the third result (Filename: `SRR5950647_1.fastq`). Add it to the same personal collection.
 
@@ -78,7 +87,7 @@ For today, here are the direct links to the two files we'll be using:
 
 ## Export your personal collection
 
-Go to your collection, and select "export" and choose "NCPI manifest format."
+Go to your collection, and select <span class="highlight_txt">export</span> and choose <span class="highlight_txt">NCPI manifest format</span>.
 
 :::info
 **What is NCPI?**
