### 2022-06-15 0abd1bf
Update CAVATICA.md
use relative fig paths
