+++ docs/Common-Fund-Tools/CFDE-Portal/iHMP/CAVATICA.md
@@ -14,17 +14,17 @@ To add data to CAVATICA, complete the following steps.
 
 Login to CAVATICA https://cavatica.sbgenomics.com/home
 
-  ![drs-1](../../images/drs-1.png)
+![drs1](https://i.imgur.com/KlGNa2Y.png)
 
 Click "Yes, I authorize" to allow CAVATICA to view your Researcher Auth Service (RAS) information
 
-![drs-2](../../images/drs-2.png)
+![drs2](https://i.imgur.com/E3LjLsr.png)
 
 Create a new project by clicking "+ Create Project". Provide a name and billing group then click "Create".  
 
 Click the "Files" tab then click "+ Add files" drop down menu select "GA4GH Data Repository Service (DRS)".  
 
-![drs-3](../../images/drs-3.png)
+![drs-3](https://i.imgur.com/zWX7W9M.png)
 
 Enter the DRS identifiers (separated by a comma, enter, or tab) and click "Import"
 * _(Follow the steps in [the pentultimate lesson](./find-export) to get 2 DRS identifiers)_
@@ -33,4 +33,4 @@ Enter the DRS identifiers (separated by a comma, enter, or tab) and click "Impor
 
 Wait for the import to finish then click "Download" to download a raw file locally confirm that it is the correct format.
 
-![drs-4](../../images/drs-4.png)
+![drs-4](https://i.imgur.com/mLcE2wZ.png)
