### 2022-06-15 e5cafec
Update CAVATICA.md
add headers and links to prereqs
