### 2022-06-16 0467c9a
Create CAVATICA-create.md
