+++ docs/Common-Fund-Tools/CFDE-Portal/iHMP/CAVATICA-create.md
@@ -1,0 +1,51 @@
+---
+layout: page
+title: Create a CAVATICA Account
+---
+
+<div class="banner"><span class="banner-text">Lesson in Development</span></div>
+
+# Create a CAVATICA Account
+
+## Step 1: Go to the CAVATICA website
+
+First, we will go to the Cavatica Registration website at
+
+[https://pgc-accounts.sbgenomics.com/auth/register](https://pgc-accounts.sbgenomics.com/auth/register).
+
+
+## Step 2: Create an Account
+
+Click  <span class="highlight_txt">Create an Account </span>.  
+
+![](https://i.imgur.com/Rcf4HwK.png)
+
+## Step 3: Choose your login system
+
+You should now have the option to continue using an eRA Commons ID, or
+by setting up a user name and password.
+
+-   If you have an eRA Commons ID, choose "login with with eRA
+    Commons" and log in using your credentials.
+-   If you don't have an eRA Commons ID, choose "Continue with email
+    and password" and then fill out the form and click
+    "Register".
+
+!!! Important
+
+    While you will be able to explore data in the Cavatica regardless of
+    which mechanism you choose, moving data from Kids First and doing
+    research requires that you have the proper authorization to view and use
+    human data, even if that data is officially "open". If you have one,
+    you should opt to login using your eRA Commons ID.
+
+
+## Step 5: Click the email link
+
+Once your registration is complete, you should get a prompt to go check
+your email:
+
+![](https://i.imgur.com/NwR4lIq.png)
+
+You will need to follow the link in that email to complete your
+registration.
