### 2022-06-16 d538a8e
Update blood-cancer-data-access.md
add outdated note
