+++ docs/Common-Fund-Tools/CFDE-Portal/index.md
@@ -1,42 +1,42 @@
 ---
 layout: page
-title: CFDE Portal Overview
+title: CFDE Search Portal Use Cases
 hide:
   - toc
 ---
 
 <div class="banner"><span class="banner-text">Lesson in Development</span></div>
 
-CFDE Portal Use Cases
-====================================================================
-
+# CFDE Search Portal Use Cases
 
 The [NIH Common Fund (CF)](https://commonfund.nih.gov) has funded a wide variety of data types and studies that are of interest to clinical and biomedical researchers, however those datasets are hosted on an equally large number of websites, with varying query systems. The [Common Fund Data Ecosystem (CFDE) Portal](https://app.nih-cfde.org) is a unified system for searching across the entire CF portfolio in a single search, and is the first step in addressing the goal of making CF data more [Find-able, Accessible, Interoperable and Reusable (FAIR)](https://www.nih-cfde.org/product/fair-cookbook/). The wide range of data types, models and formats used by Common Fund Programs are being harmonized using well-defined metadata, common controlled vocabularies using the [Crosscut Metadata Model](https://www.nih-cfde.org/product/cfde-c2m2/).
 
-Est. Time | Use Case | Lesson name | Description
---- | --- | --- | ---
-10 mins | 1 | [Cohort Selection for Blood Cancer](./Blood-Cancer/blood-cancer-portal-export.md) | Identify RNA-seq and whole genome sequencing (WGS) datasets in CFDE portal
-15 mins | 1 | [Blood Datasets from CF Program Portals](./Blood-Cancer/blood-cancer-data-access.md) | Access identified cohort in individual CF data portals
-10 mins | 2 | [Multi-omics Data Discovery for Movement-Related Disorders](./Movement-Related-Disorders/movement-disorders-portal-export.md) | Identify multiomics datasets
-10 mins | 2 | [Movement Disorders Datasets in CF Program Portals](./Movement-Related-Disorders/movement-disorders-data-access.md) | Access identified datasets in individual CF data portals
-30 mins | 3 | [Using the portal to find and export file information](./iHMP/jupyterhub.md) | Find iHMP data with DRS IDs
-30 mins | 3 | [Connect to your AWS instance on JupyterHub](./iHMP/jupyterhub.md) | Analyze whole genome sequencing data
-10 mins | 3 | [Adding data to CAVATICA with DRS IDs](./iHMP/CAVATICA.md) | Add files to CAVATICA
+The [CFDE Search Portal](https://app.nih-cfde.org/) uses the Crosscut Metadata Model (C2M2) a flexible metadata standard for describing experimental resources in biomedicine and related fields. This portal supports faceted search of metadata concepts such as anatomical location, species, and assay type, across a wide variety of datasets using a controlled vocabulary (we do not currently support protected metadata). This allows researchers to find a wide variety of data that would otherwise need to be searched individually, using varying nomenclatures. The portal only accepts C2M2 data packages from Common Fund Programs.
+
+This tutorial focuses on the Human Microbiome Project (HMP). The goal of this tutorial is to identify small FASTQ files with persistent identifiers from a longitudinal multi-comics study. 
+
+### Human Mirobiome Project Use Case
+
+Est. Time |  Lesson name | Description |
+--- | --- | --- |
+30 mins | [Using the CFDE Search Portal to Find Files](./CFDE-Portal/iHMP/find-export.md) | Find Human Microbiome Project (HMP) data with DRS IDs | 
+10 mins | [Creating a CATATICA Account](./CFDE-Portal/iHMP/CAVATICA-create.md) | Add files to CAVATICA using DRS IDs |
[…]
