### 2022-06-16 6e8454f
Update README.md
link contributing and website style guide
