### 2022-06-16 454e6cc
Update CAVATICA-drs.md
add click formatting
