+++ docs/Common-Fund-Tools/CFDE-Portal/iHMP/CAVATICA-drs.md
@@ -13,17 +13,17 @@ Login to CAVATICA https://cavatica.sbgenomics.com/home
 
 ![drs1](https://i.imgur.com/KlGNa2Y.png)
 
-Click "Yes, I authorize" to allow CAVATICA to view your Researcher Auth Service (RAS) information
+Click <span class="highlight_txt">Yes, I authorize.</span> to allow CAVATICA to view your Researcher Auth Service (RAS) information
 
 ![drs2](https://i.imgur.com/E3LjLsr.png)
 
 ## Create a Project
 
-Create a new project by clicking "+ Create Project". Provide a name and billing group then click "Create".  
+Create a new project by clicking <span class="highlight_txt">+ Create Project</span>. Provide a name and billing group then click <span class="highlight_txt">Create</span>.  
 
 ## Add files
 
-Click the "Files" tab then click "+ Add files" drop down menu select "GA4GH Data Repository Service (DRS)".  
+Click the <span class="highlight_txt">Files</span>  tab then click  <span class="highlight_txt">+ Add files</span> drop down menu select "GA4GH Data Repository Service (DRS)".  
 
 ![drs-3](https://i.imgur.com/zWX7W9M.png)
 
@@ -32,6 +32,6 @@ Enter the DRS identifiers (separated by a comma, enter, or tab) and click "Impor
 * drs://drs.hmpdacc.org/mZBm6TYnDQoS
 * drs://drs.hmpdacc.org/1CzGt5DyGwcOe
 
-Wait for the import to finish then click "Download" to download a raw file locally confirm that it is the correct format.
+Wait for the import to finish then click <span class="highlight_txt">Download</span>  to download a raw file locally confirm that it is the correct format.
 
 ![drs-4](https://i.imgur.com/mLcE2wZ.png)
