### 2022-06-16 a398b8b
Update index.md
fix typo
