### 2022-06-17 39980a3
fix intro to AWS link in dec release notes
