### 2022-06-17 205b9ad
fix my fix for December release notes
