### 2022-06-17 f01d951
formatting for commands and accessibility (remove contraction)
