### 2022-06-17 418a6d7
update outputs and fix formatting for inputs and outputs
