+++ docs/General-Tools/UNIX/navigate.md
@@ -37,8 +37,7 @@ ls
 We can see the following files:
 
 ```
-books          GTEx    MiSeq      rstudio-console.png   seattle           site-templates
-CFDE-logo.png  images  README.md  rstudio-terminal.png  site-config.toml
+books  CFDE-logo.png  GTEx  images  MiSeq  README.md  rstudio-console.png  rstudio-terminal.png  seattle
 ```
 
 If we want more information about the files, such as the date they were created and their file size, we can add "flags" `-l` for long listing format.
@@ -48,17 +47,16 @@ ls -l
 ```
 
 ```
-drwxr-xr-x 2 jovyan jovyan   4096 May 26 03:05 books
--rw-r--r-- 1 jovyan jovyan  71154 May 26 03:05 CFDE-logo.png
-drwxr-xr-x 1 jovyan jovyan   4096 May 26 03:05 GTEx
-drwxr-xr-x 2 jovyan jovyan   4096 May 26 03:05 images
-drwxr-xr-x 2 jovyan jovyan   4096 May 26 03:05 MiSeq
--rw-r--r-- 1 jovyan jovyan   2679 May 26 03:05 README.md
--rw-r--r-- 1 jovyan jovyan 239942 May 26 03:05 rstudio-console.png
--rw-r--r-- 1 jovyan jovyan 188705 May 26 03:05 rstudio-terminal.png
-drwxr-xr-x 2 jovyan jovyan   4096 May 26 03:05 seattle
--rw-r--r-- 1 jovyan jovyan    576 May 26 03:05 site-config.toml
-drwxr-xr-x 2 jovyan jovyan   4096 May 26 03:05 site-templates
+total 520
+drwxr-xr-x 2 jovyan jovyan   4096 Jun  7 04:37 books
+-rw-r--r-- 1 jovyan jovyan  71154 Jun  7 04:36 CFDE-logo.png
+drwxr-xr-x 1 jovyan jovyan   4096 Jun  7 04:36 GTEx
+drwxr-xr-x 2 jovyan jovyan   4096 Jun  7 04:37 images
+drwxr-xr-x 2 jovyan jovyan   4096 Jun  7 04:37 MiSeq
+-rw-r--r-- 1 jovyan jovyan   2679 Jun  7 04:37 README.md
+-rw-r--r-- 1 jovyan jovyan 239942 Jun  7 04:37 rstudio-console.png
+-rw-r--r-- 1 jovyan jovyan 188705 Jun  7 04:37 rstudio-terminal.png
+drwxr-xr-x 2 jovyan jovyan   4096 Jun  7 04:37 seattle
 ```
 
 Flags (sometimes called options) allow us to finely control the behavior of the command. But how did we know to add `-l` after ls? The [`ls` manual ](https://man7.org/linux/man-pages/man1/ls.1.html) describes the command and all its options in detail. Like most commands, you can type the command followed `--help` to view the manual in your terminal.
[…]
