### 2022-06-17 a56a880
fix special format navigate
