### 2022-06-17 fbc70b9
add tail references and gunzip after text
