### 2022-06-20 780ed21
Update find-export.md
fix info formatting
