### 2022-06-20 f389cd4
Update find-export.md
white space for bullets
