### 2022-06-20 fecc48b
Update find-export.md
rm bullets
