+++ docs/Common-Fund-Tools/CFDE-Portal/iHMP/find-export.md
@@ -78,9 +78,7 @@ Repeat linking to a collection with the third result (Filename: `SRR5950647_1.fa
 
 
 !!! info
-    For today, here are the direct links to the two files we'll be using:    
-    -   [file 1, `SRR5935743_1.fastq`](https://app.nih-cfde.org/chaise/record/#1/CFDE:file/nid=528892)
-    -   [file 2, `SRR5950647_1.fastq`](https://app.nih-cfde.org/chaise/record/#1/CFDE:file/nid=531342)
+    For today, here are the direct links to the two files we'll be using: [file 1, `SRR5935743_1.fastq`](https://app.nih-cfde.org/chaise/record/#1/CFDE:file/nid=528892) and [file 2, `SRR5950647_1.fastq`](https://app.nih-cfde.org/chaise/record/#1/CFDE:file/nid=531342).
 
 
 (Note, you could select any files you like, but these are small enough to work and I know what the results will be. So it's good for today's demo; I suggest trying new/different files as a Thursday exercise!)
