### 2022-06-21 0051500
Update docs/General-Tools/UNIX/navigate.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
