### 2022-06-21 e9da6d8
Update docs/General-Tools/UNIX/read.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
