### 2022-06-21 1db6459
Update docs/General-Tools/UNIX/redirect.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
