### 2022-06-21 cac5ee8
Update docs/General-Tools/UNIX/appendix.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
