### 2022-06-21 f729d35
Update docs/General-Tools/UNIX/appendix.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
