+++ docs/General-Tools/UNIX/appendix.md
@@ -105,7 +105,6 @@ done
 
 cd
 history
-history > history.txt
 
 ```
 
