### 2022-06-21 ad6c301
Update docs/General-Tools/UNIX/appendix.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
