+++ docs/General-Tools/UNIX/appendix.md
@@ -104,7 +104,6 @@ done
 
 
 cd
-history
 
 ```
 
