### 2022-06-21 2f2052d
Update docs/General-Tools/UNIX/appendix.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
