### 2022-06-21 3289ba0
Update docs/General-Tools/UNIX/appendix.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
