### 2022-06-21 d182795
Update docs/General-Tools/UNIX/redirect.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
