### 2022-06-21 e60d4b6
Update docs/General-Tools/UNIX/redirect.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
