+++ docs/General-Tools/UNIX/appendix.md
@@ -6,7 +6,8 @@ All the code from this lesson in one chunk.
 
 ```bash
 clear
-pwd
+PS1='$ '
+clear
 PS1="\w $ "
 
 ls
