### 2022-06-21 6635b68
Apply suggestions from code review
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
