### 2022-06-21 6895132
Update redirect.md
rename history file
