+++ docs/General-Tools/UNIX/appendix.md
@@ -5,37 +5,28 @@ title: Appendix
 All the code from this lesson in one chunk.
 
 ```bash
-clear
 PS1='$ '
 clear
 
-ls
-ls --help
-ls -l books
-ls -l MiSeq
-ls -lh *
+pwd
 
-cd books
-pwd
 ls
+
+ls -l
+
+ls --help
+
+cd books/
 pwd
-ls
-ls F3*
+ls -lh
 
 cd ~/books/
-
 head README.md
 
-head book.txt
-tail book.txt
-cat book.txt
-less book.txt
+head *.txt
 
-head *txt
+gunzip WizardOfOz.txt.gz
 
-wc -l *txt | sort
-
-gunzip TheWonderfulWizardofOz.txt
 cp book.txt book-copy.txt
 ls
 
@@ -47,12 +38,13 @@ mkdir -p data/results/images
 rmdir data/results/images
 rmdir data/results
 
-
 cd ~/MiSeq
 ls
 
-head -4 F3D0_S188_L001_R1_001.fastq
+head -n 4 F3D0_S188_L001_R1_001.fastq
+
 tail -4 F3D0_S188_L001_R1_001.fastq
+
 head -4 HMP_MOCK.v35.fasta 
 
 cd ../data/MiSeq/
@@ -60,6 +52,7 @@ grep CATTAG F3D0_S188_L001_R2_001.fastq
 grep CATTAG *.fastq
 
 wc F3D0_S188_L001_R2_001.fastq
+
 wc -l F3D0_S188_L001_R2_001.fastq
 
 mkdir results
@@ -67,6 +60,7 @@ grep CATTAG *.fastq > results/files-with-CATTAG.txt
 
 grep CATTAG *.fastq 
 grep CATTAG *.fastq | wc -l
+
 for file in *fastq
 do
 grep CATTAG $file | wc -l
@@ -81,6 +75,5 @@ done
 history
 
 history > history.txt
-
 ```
 
