### 2022-06-21 4af40d6
Update work.md
remove or
