+++ docs/General-Tools/UNIX/find.md
@@ -106,7 +106,7 @@ TGGGGAATATTGCACAATGGGCGAAAGCCTGATGCAGCGACGCCGCGTGAGGGATGGAGGCCTTCGGGTTGTAAACCTCT
 FASTQ and FASTA files are often used in combination to map reads to a genome or transcription. You've seen three ways to read large files. Next, we will learn how to use and manipulate the files.  
 
 
-## Wildcards
+### Wildcards
 
 Sometimes you know a file or directory exists, but you can't find it. Sometimes you want to find many files with similar properties. This is where the wildcard (`*`) comes in handy.  What do the following commands do?
 
@@ -119,8 +119,8 @@ Sometimes you know a file or directory exists, but you can't find it. Sometimes 
 !!! spoiler
 
      1. `ls *` lists files in the working directory and 1 level down. 
-1. `ls MiSeq/F3D*` lists files in the data/MiSeq directory that start with "F3D".
-1. `ls MiSeq/*fasta` lists files in the data/MiSeq directory that end with "fasta".
+     1. `ls MiSeq/F3D*` lists files in the data/MiSeq directory that start with "F3D".
+     1. `ls MiSeq/*fasta` lists files in the data/MiSeq directory that end with "fasta".
 
 
 
@@ -133,7 +133,6 @@ The `MiSeq/` directory contains many of the sequence files ending in`.fastq`. We
 Let's say you'd like to find the sequence `CATTAG` in your MiSeq files. We can use the function `grep` to search for  `CATTAG` in one or all of the fastq files located in our current working directory.
 
 ```
-cd ../data/MiSeq/
 grep CATTAG F3D0_S188_L001_R2_001.fastq
 grep CATTAG *.fastq
 ```
@@ -149,20 +148,92 @@ Use `grep --help` to search for `grep` options related to line number.
 `grep -n [filename]`` will print the line number.
 
 
-### `wc`
+In addition to searching for nucleotide sequences, you may want to search for information in the first line of a .fastq or .fasta file. The `^` (shift + 6) can be used to specify "the beginning of the line".
 
-How many lines are in the file total? The word count (`wc`) command will give us that answer.
+```
+grep "^>" *fasta
 
-```bash
-wc F3D0_S188_L001_R2_001.fastq
 ```
 
-By default, `wc` prints the characters, words, and lines in a file. To extract just the line numbers, we use `wc -l`. We will use this more later. 
+This will print the name associated with a given sequence in the searched files. In this case, there is only one fasta file, so the name is not printed. 
 
-
-```bash
-wc -l F3D0_S188_L001_R2_001.fastq
 ```
+>A.baumannii.1
+>A.odontolyticus.1
+>B.cereus.1
+...
+>S.agalactiae.1
+>S.mutans.1
+>S.pneumoniae.1
+```
+
+We can also print the line before or after the line that matches a pattern with `-B 1` `-A 1`, respectively.
+
+```
+grep -A 1 "^>" *fasta
+
+```
+
+This will print the name and sequence for every entry. The first is shown here.
+
+```
+>A.baumannii.1
[…]
