### 2022-06-21 895165b
Update UNIX.md
first pass syncing hackmd source file with lesson material
