+++ docs/General-Tools/UNIX/redirect.md
@@ -4,6 +4,33 @@ title: Redirect
 
 In this section, we will explore redirecting, appending, piping, and looping with `>`, `>>`,  `|`, and `for`.
 
+If you completed the last challenge, you saw that the **images/** directory contains a file called **MiSeq-readcount-Mothur.png**. This image is a screenshot from the [Mauther software tutorial](https://mothur.org/wiki/miseq_sop/) showing the count or number of reads for each sample. 
+
+![](https://hackmd.io/_uploads/HyrJqfBTF.png)
+
+To see if our data matches theirs, we can count the number of lines in the .fastq files with the UNIX command `wc`. This will print by default the number of characters, words, and lines in a file. We can ask for just the number of lines with the `-l`  option. 
+
+```
+wc -l *fastq
+```
+
+This gives something like:
+
+```
+31172 F3D0_S188_L001_R1_001.fastq
+31172 F3D0_S188_L001_R2_001.fastq
+23832 F3D141_S207_L001_R1_001.fastq
+23832 F3D141_S207_L001_R2_001.fastq
+ ... 
+28280 F3D9_S197_L001_R1_001.fastq
+28280 F3D9_S197_L001_R2_001.fastq
+19116 Mock_S280_L001_R1_001.fastq
+19116 Mock_S280_L001_R2_001.fastq
+1218880 total 
+```
+
+However, this number is too large. In fact, it is 4 times larger than the number of reads. To capture just the number of reads, we can first use `grep` followed `wc`.
+
 By default, many UNIX commands like `cat` send output to something called
 standard out, or "stdout". This is a catch-all phrase for "the basic
 place we send regular output." (There's also standard error, or "stderr",
@@ -15,33 +42,60 @@ stdout output, and if you work with UNIX a lot, you'll see characters
 like the `>` (redirect), `>>` (append) , and `|` (pipe) thrown around. These
 are redirection commands that say, respectively, "send stdout to a new
 file", "append stdout to an existing file", and "send stdout from one
-program to another program's stdin."
+program to another program's stdin. If you know you want to save an output file, you can use the redirect symbol `>`. 
+Note, if you want to save a file in a different directory, that directory must exist.
 
-If you know you want to save an output file, you can use the redirect symbol `>`. Note, if you want to save a file in a different directory, that directory must exist.
-
-```bash
-mkdir results
-grep CATTAG *.fastq > results/files-with-CATTAG.txt
-```
 
 ### `|`
 
-Let's pipe (`|`) the output of a grep search to the command head to only show the first 10 lines. This is very handy when you only want to view a snapshot of the data. 
+Let's now use grep to match the first line, which starts with "@M00967", of all the R1 files then pipe the output to wc and count the number of liens. 
 
-
-```bash
-grep CATTAG *.fastq 
-grep CATTAG *.fastq | wc -l
+```
+head -n 1 *.fatsq
+grep "^@M00967" *R1*.fastq | wc -l
 ```
 
-The pipe is a useful way to combine multiple commands on line line into a pipe. 
+The answer, 152883, matches the authors. Nice. Also, we just scanned many large files very quickly to confirm a finding. 
+
+```
+152883
+
+```
+
+You probably don't want to read all the lines that were matched, but piping the output to head is a nice way to view the first 10 lines. 
+
+```
[…]
