### 2022-06-21 a90802b
Update UNIX.md
upate redirect section
