### 2022-06-21 63961a1
hint wording for ebook challenge
