+++ hackmd/UNIX.md
@@ -557,7 +557,7 @@ GATTTGGGGTTCAAAGCAGTATCGATCAAATAGTAAATCCATTTGTTCAACTCACAGTTT
 ```
 :::
 
-We can use the `cat` command to print fastq files to the screen, but thousands of lines of text would crowd your screen. Instead, let's use the `head` command to view the first 8 lines file. You can copy the file name and paste it into the console or you can type and use tab complete to pick a particular file. 
+We can use the `cat` command to print fastq files to the screen, but thousands of lines of text would crowd your screen. Instead, we will use the `head` command to view the first 8 lines of the file. You can copy the file name below and paste it into the console or you can type and use tab complete to pick a particular file.
 
 ```
 head -n 4 F3D0_S188_L001_R1_001.fastq
@@ -621,17 +621,17 @@ TAGGGAATCTTCGGCAATGGACGGAAGTCTGACCGAGCAACGCCGCGTGAGTGAAGAAGGTTTTCGGATCGTAAAGCTCT
 
 ### Wildcards 
 
-Sometimes you know a file or directory exists, but you can't find it. Sometimes you want to find many files with similar properties. This is where the wildcard (`*`) comes in handy. What do the following commands do?
+Sometimes you know a file or directory exists, but you can not find it. Sometimes you want to find many files with similar properties. This is where the wildcard (`*`) comes in handy. What do the following commands do?
 
 1. `ls *` 
 1. `ls F3D*`
-1. `ls *fasta`
+1. `ls *.fasta`
  
 :::spoiler
 
 1. `ls *` lists files in the working directory and 1 level down. 
 1. `ls MiSeq/F3D*` lists files in the data/MiSeq directory that start with "F3D".
-1. `ls MiSeq/*fasta` lists files in the data/MiSeq directory that end with "fasta".
+1. `ls MiSeq/*.fasta` lists files in the data/MiSeq directory that end with "fasta".
 
 :::
 
@@ -641,7 +641,7 @@ A lot of the time we want to know if a file contains what we expect. A useful th
 
 The `MiSeq/` directory contains many of the sequence files ending in`.fastq`. We expect these files to contain information in a particular format throughout the file with four lines of information for each sequence string. Looking through a million-line file using less will take a long time. Rather than manually looking at the whole file, we can print only a portion of the file's contents to standard output. 
 
-Let's say you'd like to find the sequence `CATTAG` in your MiSeq files. We can use the function `grep` to search for  `CATTAG` in one or all of the fastq files located in our current working directory.
+Let's say you would like to find the sequence `CATTAG` in your MiSeq files. We can use the function `grep` to search for  `CATTAG` in one or all of the fastq files located in our current working directory.
 
 ```
 grep CATTAG F3D0_S188_L001_R2_001.fastq
@@ -660,7 +660,7 @@ Use `grep --help` to search for `grep` options related to line number.
 In addition to searching for nucleotide sequences, you may want to search for information in the first line of a .fastq or .fasta file. The `^` (shift + 6) can be used to specify "the beginning of the line".
 
 ```
-grep "^>" *fasta
+grep "^>" *.fasta
 
 ```
 
@@ -679,7 +679,7 @@ This will print the name associated with a given sequence in the searched files.
 We can also print the line before or after the line that matches a pattern with `-B 1` `-A 1`, respectively.
 
 ```
-grep -A 1 "^>" *fasta
+grep -A 1 "^>" *.fasta
 
 ```
 
@@ -693,16 +693,16 @@ TGGGGAATATTGGACAATGGGGGGAACCCTGATCCAGCCATGCCGCGTGTGTGAAGAAGGCCTTATGGTTGTAAAGCACT
 
 ### `find`
 
[…]
