### 2022-06-22 f852f60
Update docs/General-Tools/R-for-RNA-Seq/import.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
