### 2022-06-22 fa96be8
Update docs/General-Tools/R-for-RNA-Seq/intro.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
