### 2022-06-22 d24702b
Update docs/General-Tools/R-for-RNA-Seq/import.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
