### 2022-06-22 c7925b6
Update docs/General-Tools/R-for-RNA-Seq/visualize.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
