### 2022-06-22 0612e37
Update docs/General-Tools/R-for-RNA-Seq/wrangle.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
