### 2022-06-22 63156a0
Update docs/General-Tools/R-for-RNA-Seq/wrangle.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
