### 2022-06-22 e809d2f
Update docs/General-Tools/R-for-RNA-Seq/wrangle.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
