### 2022-06-22 106b9fc
Update docs/General-Tools/R-for-RNA-Seq/wrangle.md
Co-authored-by: jessicalumian <jessica.mizzi@gmail.com>
