+++ docs/Website-Style-Guide/7ReleasePlan.md
@@ -1,58 +1,46 @@
 # CFDE Training Website Release Plan
 
+
 This is the release plan for the [CFDE training website](https://cfde-training-and-engagement.readthedocs-hosted.com/en/latest/).
 This plan has sections to address the different release management topics:
 
--  Definition of release numbering and naming convention
--  Definition of expected release frequency
+-  Expected release frequency
+-  Release numbering and naming convention
 -  Locations for release documentation, repo etc
 -  Roles and responsibilities for the releases
--  Definition of steps in a release cycle
+-  Release standard operating procedure (SOP)
 
 ## Release Dates
 
-Timeline for release 2020-2021
+Our current release dates are set to coincide with NIH deliverable dates which are due on the 15th of every month. Previous releases can be viewed on [GitHub](https://github.com/nih-cfde/training-and-engagement/releases) or on the training website [Release page](https://training.nih-cfde.org/en/latest/Release-Notes/).
 
 
-| Release # | Date | Month     | Year | Version tag |
-| --------- | ---- | --------- | ---- | ----------- |
-| 1         | 15   | August    | 2020 | 2020.08     |
-| 2         | 15   | October   | 2020 | 2020.10     |
-| 3         | 18   | December  | 2020 | 2020.12     |
-| 4         | 28   | February  | 2021 | 2021.02     |
-| 5         | 31   | May       | 2021 | 2021.05     |
-| 6         | 31   | August    | 2021 | 2021.08     |
-| 7         | 30   | November  | 2021 | 2021.11     |
+## Release Title and Tag Format
+
+Given the montly cycle, releases are named with the month and year. 
+
+**Title:** Month Year (e.g. October 2020)  
+**Tag:** Year.<area>Mo (e.g. 2020.10)
 
 
-Our current release dates are set to coincide with NIH deliverable dates.
+## Release Checklists
 
 
-#### Labels Format
-Single event based label
-Name: *Training-Release*
+The [`.github` directory](https://github.com/nih-cfde/training-and-engagement/tree/dev/.github) contains actions, workflows, and templates with checklists for to ensure that tutorials and new website pages are error-free when released to the public. These inlude auto-assigning reviewers to PRs against `dev` and `stable`, adding a reviewer checklist to PRs, auto-building the mkdocs-based website, and drafting a GitHub relase. 
 
-Month based label
-Name: release.month(short form)-release.year
-For example, *Oct-2020*.
+Training materials should follow the formatting, organization, and layout details described in the [training website style guide](../Website-Style-Guide/0index.md). 
 
-## Release Checklist
+### Submitting New Website Content
 
-A series of checklists will be implemented to ensure error free working tutorials are released to the public.
+New content for the website should be submitted as a pull request against the `dev` branch following the steps below.  
 
-### Tutorial Format
-
-Individual tutorials will follow the formatting, organization, and layout details described in the [training website style guide](../Website-Style-Guide/0index.md).
-
-### Tutorial PR Process
-
-Individual branches are used to create PRs to merge into `dev`. Good practices for creating PR branches are:
-- create a new branch off `dev`
-- name the branch with included feature/fix/content along with PR author's name. e.g. scanchi-mime-new or marisa-sphinx-javascript or abhijna_kf_edits
-- keep the PR branch up to date with `dev` by pulling latest merged changes. *Note on the GitHub interface for the PR you will get an indication to update your branch which essentially merges latest changes in `dev` to your branch*
-- After the PR has been successfully merged, delete the branch in your remote and local repos
-
[…]
