+++ docs/Website-Style-Guide/index.md
@@ -8,16 +8,17 @@ This is a style guide for content on the CFDE [training website](https://cfde-tr
 Time | Section | About
 --- | --- | ---
 5 mins | 1. [Website navigation](./1WebsiteNavStyle.md) | How to navigate the website?
-5 mins |2. [Website style guide](./1WebsiteNavStyle.md) | How is the website formatted?
-10 mins | 3. [General tutorial text style guide](./2GenTutorialStyle.md) | How are CFDE tutorials formatted?
-10 mins | 4. [Tutorial components](./3TutorialComponents.md) | How are tutorial contents organized?
-10 mins | 5. [Tutorial recordings style guide](./4RecordingStyleGuide.md) | How to create vidlets or screencasts?
-10 mins | 6. [Pangeo binder style guide](./5PangeoBinderGuide.md)| How to create Pangeo binders?
-5 mins | 7. [Practice repo style guide](./6PracticeGithubRepos.md) | How to create a practice Github repo?
+10 mins | 2. [General tutorial text style guide](./2GenTutorialStyle.md) | How are CFDE tutorials formatted?
+10 mins | 3. [Tutorial components](./3TutorialComponents.md) | How are tutorial contents organized?
+10 mins | 4. [Tutorial recordings style guide](./4RecordingStyleGuide.md) | How to create vidlets or screencasts?
+10 mins | 5. [Pangeo binder style guide](./5PangeoBinderGuide.md)| How to create Pangeo binders?
+5 mins | 6. [Practice repo style guide](./6PracticeGithubRepos.md) | How to create a practice Github repo?
+10 mins | 7. [Release Plan](./7ReleasePlan.md)| What is the tutorial review process? How often are new lessons added to the website?
+5 mins | 8. [Template file](./8TutorialTemplate.md) | How do I format my files?
 
 === "Resources"
 
-    - [Get started with new tutorials using the tutorial template markdown doc](https://github.com/nih-cfde/training-and-engagement/blob/dev/docs/CFDE-Internal-Training/Website-Style-Guide/tutorial_template_docs/TutorialTemplate.md)
+    - [Get started with new tutorials using the tutorial template markdown doc](https://github.com/nih-cfde/training-and-engagement/blob/dev/docs/CFDE-Internal-Training/Website-Style-Guide/8TutorialTemplate.md)
     - [Powerpoint slide template](https://drive.google.com/drive/u/0/folders/14dOaf7-G4k7rCw5mL2Q5jdRWXrO0Y5i-)
 
 === "Prerequisites"
