### 2022-06-22 b7ed828
change aws5 to aws3 (screen in feb 2021 update)
