### 2022-06-22 88587a3
second attempt to fix link to video (index page)
(.) and (./index.md) not working
