### 2022-06-22 c55169c
remove video link and replace with text
