### 2022-06-23 8d8c1e7
Update index.md
add prereqs
