### 2022-07-10 #167 @abhijna
Hi @willingc, here is the:
- GitHub release tag for the branch containing the aws manuscript: https://github.com/nih-cfde/training-and-engagement/releases/tag/2022.07-aws-jose
- Zenodo archive here: https://zenodo.org/record/6816045#.Yss_yuzMIUv

Please let me know if you need anything else!
