### 2022-08-11 #167 @labarba
I see that this submission has been flagged for publication. However, the reviewers did not complete their review checklist. Can I have confirmation from @giuliaguidi, @manics that they are ready to check off all items in their list?
