### 2022-08-12 #167 @giuliaguidi
Yes

Giulia

On Thu, Aug 11, 2022 at 3:54 PM Lorena A. Barba ***@***.***>
wrote:

> I see that this submission has been flagged for publication. However, the
> reviewers did not complete their review checklist. Can I have confirmation
> from @giuliaguidi <https://github.com/giuliaguidi>, @manics
> <https://github.com/manics> that they are ready to check off all items in
> their list?
>
