### 2022-08-12 #167 @manics
The version mentioned in this issue `1.0.0` doesn't match the repository versions at https://github.com/nih-cfde/training-and-engagement/releases
but presumably that's easy to fix. Everything else is fine!
