### 2022-08-12 #167 @labarba
Thank you, @manics. We will update the version once all is ready for publication.

For a complete record, could I ask both reviewers to check off the remaining items in their review checklist?
