### 2022-08-12 #167 @labarba
@abhijna  – Can you confirm that you would like the _version_ to be marked as [2022.07-aws-jose](https://github.com/nih-cfde/training-and-engagement/releases/tag/2022.07-aws-jose)?
