### 2022-08-12 #167 @giuliaguidi
Hi, I’m traveling today I’ll make sure to check off the list by Monday.

Giulia

On Fri, 12 Aug 2022 at 08:07, Abhijna Parigi ***@***.***>
wrote:

> @abhijna <https://github.com/abhijna> – Can you confirm that you would
> like the *version* to be marked as 2022.07-aws-jose
> <https://github.com/nih-cfde/training-and-engagement/releases/tag/2022.07-aws-jose>
> ?
>
> Yes, I can confirm this is correct!
>
