### 2022-08-26 #167 @labarba
Congratulations, @abhijna, your JOSE paper is now published! 🚀

Huge thanks to our Editor: @willingc, and Reviewers: @giuliaguidi, @manics — your contributions make this adventure in publishing possible, and we are all grateful for it 🙏
